### 2019-08-16 #1652 @whedon
[PRE REVIEW]: UncertainData.jl: a Julia package for working with measurements and datasets with uncertainties
**Submitting author:** @kahaaga (<a href="http://orcid.org/0000-0001-6880-8725">Kristian Agasøster Haaga</a>)
**Repository:** <a href="https://github.com/kahaaga/UncertainData.jl" target ="_blank">https://github.com/kahaaga/UncertainData.jl</a>
**Version:** v0.2.1
