### 2019-08-16 #1652 @danielskatz
👋  @kahaaga - we can't proceed until you add an OSI-compliant license to the repo.  Also, please change manuscript.md to paper.md
