### 2019-08-16 #1652 @kahaaga
Done, @danielskatz! An MIT-license has been added, and the manuscript file has been renamed to paper.md.

I also had a look at the list of potential reviewers and found some Julia programmers with some relevant backgrounds (not sure if I should tag them here, so here are their github handles):
- Datseris
- dmbates
- matbesancon
- juliohm
