### 2019-08-16 #1652 @danielskatz
You’ll need to fix the affiliations. See the same paper please. Then add a new comment to this issue that is ```@whedon generate pdf``` to try again
