### 2019-08-17 489ecc7
Fix affiliations in paper.md
