### 2019-08-17 7052bdf
Add travis badge.
