+++ README.md
@@ -1,3 +1,5 @@
+[![Build Status](https://travis-ci.com/kahaaga/UncertainData.jl.svg?branch=master)](https://travis-ci.com/kahaaga/UncertainData.jl)
+
 # UncertainData.jl
 A Julia package for dealing with data values with associated uncertainties and
 datasets consisting of uncertain values.
@@ -11,7 +13,6 @@ custom user-defined constraints.
 
 3. Provide a framework for robust computation of ensemble statistics for
 uncertain data.
-
 
 Please check out the
 [documentation](https://kahaaga.github.io/UncertainData.jl/dev) for more
