+++ paper/paper.md
@@ -7,7 +7,7 @@ tags:
 authors:
   - name: Kristian Agasøster Haaga
     orcid: 0000-0001-6880-8725
-    affiliation: "1, 2", "3"
+    affiliation: "1", "2", "3"
 affiliations:
   - name: Department of Earth Science, University of Bergen, Bergen, Norway
     index: 1 
