### 2019-08-17 d20db5c
Error on line 8 might be a problem with norwegian letters
Try with shortened middle name.
