### 2019-08-17 596eb3c
Try unicode for `ø`
