### 2019-08-17 f8a175d
Capitalization on letters in bib file
