+++ paper/paper.bib
@@ -17,7 +17,7 @@
 }
 
 @article{Vasskog:2019,
-  title={Evidence of early deglaciation (18 000 cal a bp) and a postglacial relative sea-level curve from southern Karm{\o}y, south-west Norway},
+  title={Evidence of early deglaciation (18 000 cal a bp) and a postglacial relative sea-level curve from southern {K}arm{\o}y, south-west {N}orway},
   author={Vasskog, Kristian and Svendsen, John-Inge and Mangerud, Jan and Agas{\o}ster Haaga, Kristian and Svean, Arve and Lunnan, Eva Maria},
   journal={Journal of Quaternary Science},
   year={2019},
