### 2019-08-17 de15fee
Try \{o} syntax to get whedon to parse middle name with ø
