+++ paper/paper.md
@@ -5,7 +5,7 @@ tags:
   - uncertainty
   - measurements
 authors:
-  - name: Kristian Agas\u00F8ter Haaga
+  - name: Kristian Agas\{o}ster Haaga
     orcid: 0000-0001-6880-8725
     affiliation: "1, 2, 3"
 affiliations:
