### 2019-08-17 891daf3
Change back to `ø` again, error might be unrelated
The affiliation numbering was formatted incorrectly in the same commit, so there might actually not be a problem here
