### 2019-08-17 #1652 @kahaaga
Affiliations are fixed!

There was some confusion about a parsing error on line 8 which I thought to be due to the Norwegian letter `ø` in my middle name. However, this which turned out to be due to incorrect formatting of the `affiliation` field, not an error in parsing the letter `ø`. The pdf is now rendered correctly.
