### 2019-08-17 866f532
minor punctuation changes
