### 2019-08-17 #1652 @danielskatz
Thanks! Also, sorry for the perhaps confusing message I sent - it was meant to say "sample" paper but ended up saying "same" paper, as I was typing on my phone on a plane that was getting ready to leave...

Can you also merge https://github.com/kahaaga/UncertainData.jl/pull/58 to fix some small punctuation issues?
