### 2019-08-17 #1652 @kahaaga
No problem, it was obvious enough that I had made a mistake there, so no confusion about that!

I've merged your punctuation fixes. I'm also merging the `joss_paper` branch with the final fixes in `paper.md` to master. Once the CI tests pass in ten minutes or so, I'll get whedon to generate a new pdf.
