### 2019-08-17 #1652 @kahaaga
Hehe, no problem! I wasn't expecting to get a response so swiftly - thanks for the quick follow up.

Anyways, here we go: The pdf looks good, and everything is ready from my side.
