### 2019-08-17 6768378
Added link to HypothesisTests.jl and re-worded
