### 2019-08-17 #1652 @kahaaga
I added an additional reference to a Julia package pending reviewer assignment.
