### 2019-08-17 #1652 @danielskatz
Thanks, but before that, we need to assign an editor who will take on finding reviewers and managing the review process.
