### 2019-08-18 e8beeef
Fixed resampling of vectors of uncertain vals + added resample_elwise
