### 2019-08-18 1d6b184
Fix doc string
