+++ src/resampling/uncertain_dataset/resample_uncertaindataset_index.jl
@@ -12,7 +12,7 @@ function resample(uvd::UncertainIndexDataset)
 end
 
 """
-	resample(uvd::UncertainIndexDataset)
+	resample(uvd::UncertainIndexDataset, n::Int)
 
 Draw `n` realisations of an `UncertainIndexDataset` according to the
 distributions of the `UncertainValue`s comprising it.
