### 2019-08-18 27bcc7d
resample_elwise for AbstractUncertainValueDataset subtypes
