### 2019-08-19 5a1dc4b
Resample AbstractUncertainValueDataset with constraints
