### 2019-08-19 #1652 @oliviaguest
Hi @dmbates and @lindemann09, would you be interested and able to review this JOSS submission? 😄
