### 2019-08-19 #1652 @lindemann09
I don't know much about Julia. I'm afraid I do not have the required expertise to review this submission.
