### 2019-08-19 49e5b97
Add tests for using vectors of constraints.
