### 2019-08-19 fa63ea2
Restructure menu and rename menu elements in docs
