### 2019-08-19 104f6e9
Add tests for resample_elwise(::AbstractUncertainValueDataset)
