### 2019-08-19 ee4110b
Update changelog for v0.2.2
