### 2019-08-19 aaed847
Removed checklist from changelog
