+++ docs/src/changelog.md
@@ -1,21 +1,29 @@
 ## UncertainData.jl v0.2.2
 
 ### New functionality and syntax changes 
+
 #### Resampling vectors consisting of uncertain values (done in #61)
 
-- [x] `resample(uvals::Vector{AbstractUncertainValue}, n::Int)` is now interpreted as "treat `uvals` as a dataset and sample it `n` times". Thus, it now behaves as `resample(AbstractUncertainDataset, n::Int)`, returning `n` vectors of length `length(uvals)`, where the i-th element is a unique draw of `uvals[i]`.
+- `resample(uvals::Vector{AbstractUncertainValue}, n::Int)` is now interpreted as "treat 
+    `uvals` as a dataset and sample it `n` times". Thus, it now behaves as 
+    `resample(AbstractUncertainDataset, n::Int)`, returning `n` vectors of length 
+    `length(uvals)`, where the i-th element is a unique draw of `uvals[i]`.
 
-- [x] `resample_elwise(uvals::Vector{AbstractUncertainValue}, n::Int)` takes over the role as "sample `uvals` element-wise and `n` times for each element". Returns a vector of length `length(uvals)`, where the i-th element is a `n`-element vector of unique draws of `uvals[i]`.
+- `resample_elwise(uvals::Vector{AbstractUncertainValue}, n::Int)` takes over the role as 
+    "sample `uvals` element-wise and `n` times for each element". Returns a vector of 
+    length `length(uvals)`, where the i-th element is a `n`-element vector of unique draws 
+    of `uvals[i]`.
 
 #### Resampling with subtypes of `AbstractUncertainValueDataset`
 
-Currently, this affects the generic `UncertainDataset`s, as well as the specialized `UncertainIndexDataset`s and `UncertainValueDataset`s.
+Currently, this affects the generic `UncertainDataset`s, as well as the specialized 
+`UncertainIndexDataset`s and `UncertainValueDataset`s.
 
-- [x] `resample_elwise(uvd::AbstractUncertainValueDataset, n::Int)` is now interpreted as 
+- `resample_elwise(uvd::AbstractUncertainValueDataset, n::Int)` is now interpreted as 
     "draw `n` realisations of each value in `uvd`". Returns a vector of length `length(uvals)` 
     where the i-th element is a `n`-element vector of unique draws of `uvals[i]`. This works 
     for `UncertainDataset`s, `UncertainIndexDataset`s, and `UncertainValueDataset`s. 
-- [x] `resample_elwise(uvd::AbstractUncertainValueDataset, constraint::Union{SamplingConstraint, Vector{SamplingConstraint}}, n::Int)` 
+- `resample_elwise(uvd::AbstractUncertainValueDataset, constraint::Union{SamplingConstraint, Vector{SamplingConstraint}}, n::Int)` 
     is now interpreted as "draw `n` realisations of each value in `uvd`, subjecting each value 
     in `uvd` to some sampling `constraint`(s) during resampling". Returns a vector of 
     length `length(uvals)` where the i-th element is a `n`-element vector of unique draws 
@@ -24,13 +32,16 @@ Currently, this affects the generic `UncertainDataset`s, as well as the speciali
 
 ### Bug fixes
 
-- [x] Removed extra blank line from print method for `AbstractUncertainPopulation`.
+- Removed extra blank line from print method for `AbstractUncertainPopulation`.
 
 ## UncertainData.jl v0.2.1
 
 ### New functionality
 
[…]
