### 2019-08-20 cfc0707
Validate input for TruncateQuantiles
