### 2019-08-20 5781fe1
Validate input for TruncateRange
