### 2019-08-20 247964d
Input validation for TruncateStd
