### 2019-08-20 45a49aa
Simplify TruncateQuantiles validation
