### 2019-08-20 25f39c9
Add sampling constraints tests
