### 2019-08-20 0815aa6
resampling min-max validation
