### 2019-08-20 #1652 @ahwillia
Yes - I have a few deadlines and am wrapping up summer travel, so in 2
weeks I'll have some time.

On Tue, Aug 20, 2019 at 12:49 PM Olivia Guest <notifications@github.com>
wrote:

> @ahwillia <https://github.com/ahwillia> so in about 2 weeks you think you
> can get started?
>
