### 2019-09-07 #1666 @dmbates
I have finished the review of this paper and it looks very good to me.  It is a pleasure to review a paper and software like this that has been so carefully created.  My only comment is that the author has included the top-level `Manifest.toml` file in the repository and I believe it is recommended not to do that.  In other words, that name should be added to the `.gitignore` file.
