### 2019-09-07 #1666 @kahaaga
Thank you for the encouraging feedback, @dmbates.

I've added `Manifest.toml` to the `.gitignore` file. These changes are now implemented on the master branch and will be part of the next release.
