### 2019-09-07 #1666 @dmbates
Oh, one other thing I forgot to mention.  I left the check box for version matches unchecked because the repository version is now ahead of the version given in the paper.  I assume that the way to resolve this is to change the paper but I leave it to those who have more knowledge of the system to decide.
