### 2019-09-07 #1666 @labarba
Once the review is finished, and all modifications done to the software and the paper, you ask the authors to make a new tagged release (and archive in Zenodo). You will then run the command:

`@whedon set <v.x.x> as version`

and the version number will be updated in the header comment here, and in the margin decorators of the first page of the paper in subsequent compilations.
