### 2019-09-07 #1666 @labarba
See [Changing the software version](https://joss.readthedocs.io/en/latest/whedon.html?highlight=version#changing-the-software-version)
