### 2019-09-10 #1666 @ahwillia
I just want to echo all of @dmbates's positive comments. This is a very well done project and I think it should be accepted. Please treat the comments below as friendly suggestions and pointers for youre future work.

I had one question when looking through the documentation. How do functions like `cov(x, y)` and `cor(x, y)` work if `resample` is called independently on the `x` and `y` variables? It would seem like you'd need to specify a multivariate distribution (e.g. a bivariate Normal distribution) that specifies both `x` and `y`. Then you could call `resample` once, and compute the correlation.

This package also takes a frequentist perspective. It could be nice to think about and maybe include a discussion in the documentation about how Bayesian statisticians would handle this problem. Their response would be to treat the data as fixed, and instead create a probabilistic model that captures the uncertainty. There are advantages/disadvantages associated with each approach. It's possible that the Bayesian approach is more computationally efficient on large-scale data—if I have a very large, uncertain dataset that takes a long time to fit, I'd need to resample my data many times and re-fit a very expensive model. In some cases the Bayesian model fitting can be faster, but is also typically very complicated and computationally intensive.

In case there is interest, here are some references on large-scale Bayesian model fitting:

- https://arxiv.org/abs/1401.0118
- https://arxiv.org/abs/1603.00788
- https://arxiv.org/abs/1610.02287

I've checked all my boxes other than the release version number (as per above discussion).
