### 2019-09-12 1d7d258
Resample uncertain vectors, mapping each element to a unique sampling constraint
