### 2019-09-12 356f9be
Resample uncertain vectors
