### 2019-09-12 754a3aa
Restructure and fix some bugs with UncertainIndexDataset sampling
