### 2019-09-12 662ab8c
Update changelog for v0.3.0
