### 2019-09-12 191a24d
Resample vectors of uncertain values without constraints
