### 2019-09-12 5bc230d
resample vectors of uncertain values, multiple draws
