+++ docs/src/changelog.md
@@ -1,6 +1,18 @@
 
 # Changelog
 
+## UncertainData.jl v0.3.1
+
+### New functionality
+
+- Resample vectors of uncertain values without constraints. Syntax:
+    1. `resample(::Vector{<:AbstractUncertainValue}` for single draws.
+    2. `resample(::Vector{<:AbstractUncertainValue}, ::Int}` for multiple draws.
+
+- Resample vectors of uncertain values with constraint(s) multiple times. Syntax:
+    1. `resample(::Vector{<:AbstractUncertainValue}, ::Union{SamplingConstraint, Vector{<:SamplingConstraint}}` for single draws.
+    2. `resample(::Vector{<:AbstractUncertainValue}, ::Union{SamplingConstraint, Vector{<:SamplingConstraint}}, ::Int` for multiple draws.
+
 ## UncertainData.jl v0.3.0
 
 ### New functionality
