### 2019-09-12 125653d
AbstractUncertainDataResampling
Introduce an abstract resampling type for this package pending the implementation of `AbstractResampling` in StatsBase.jl
