+++ src/resampling/resampling_schemes/AbstractUncertainDataResampling.jl
@@ -1,0 +1,9 @@
+# TODO: inherit from StatsBase.AbstractResampling when implemented there
+"""
+    AbstractUncertainDataResampling
+
+An abstract type for all resampling schemes in this package.
+"""
+abstract type AbstractUncertainDataResampling end
+
+export AbstractUncertainDataResampling
