### 2019-09-12 b50ad6c
Add `ConstrainedResampling`
