+++ docs/src/changelog.md
@@ -1,9 +1,14 @@
 
 # Changelog
 
-## UncertainData.jl v0.3.1
+## UncertainData.jl v0.4.0
 
 ### New functionality
+
+- Introduce an abstract resampling type `AbstractUncertainDataResampling` for this 
+    package pending the implementation of `AbstractResampling` in StatsBase.jl.
+
+- Added `ConstrainedResampling` resampling scheme.
 
 - Resample vectors of uncertain values without constraints. Syntax:
     1. `resample(::Vector{<:AbstractUncertainValue}` for single draws.
