### 2019-09-20 28b6518
Resampling a number just returns the number
