### 2019-09-20 #1666 @kahaaga
Thanks for the nice feedback and suggestions, @ahwillia!

## How `cor(x,y) and related functions work`

For the `cov(x, y)` and `cor(x, y)` functions, the approach is to always call `resample` independently on `x` and `y`, because they might follow different distributions (hence, not necessarily normal distributions). The general approach (also for `cor(x, y)`) is thus as follows.

Assume I have two uncertain datasets `x` and `y` of length `n`, where element of the uncertain dataset is some distribution. Each element of `x`, and each element of `y` may be a regular numeric value with uncertainty, a normal distribution, an empirical distribution, a KDE to some distribution, a weighted population, or some new uncertain data type I choose to define, as long as I have a well-defined way of drawing random numbers from it.

To compute the correlation between `x` and `y`, I first resample `x` according to the distributions furnishing it, and then resample `y` according to the distributions furnishing it. That gives me one realization/draw for `x` and one realization/draw for `y`, both of length `n`. The correlation between those two draws is then computed, yielding a single value for the correlation statistic. To get a distribution on the statistic, I just repeat that process many times.

This way, I can compute *any* statistic that accepts numerical vectors as inputs from uncertain data of *any kind*. For example, in the latest version of [CausalityTools.jl](https://github.com/kahaaga/CausalityTools.jl),
I use this approach to allow the user to compute various [causality statistics from uncertain time series data](https://kahaaga.github.io/CausalityTools.jl/dev/causalitytests/causality_from_time_series/).

I purposely went for this simplest possible approach. If your data is well-described by theoretical distributions, then you have the option to represent them as such. But if your data are not, then you still can handle uncertainties, because the resampling approach works where analytical expressions are hard to derive. In fact, the user already has the option to create a probabilistic model to represent their uncertainties, both using kernel density estimation and using model fitting. This is described in the documentation [here](
https://kahaaga.github.io/UncertainData.jl/dev/uncertain_values/uncertainvalues_fitted/).

## Some more details on model fitting

 If the user has some empirical distributions, they can use basic model fitting as such:

```julia
using Distributions, UncertainData

# Create a normal distribution
d = Uniform()

# Draw a 1000-point sample from the distribution.
some_sample = rand(d, 1000)

# Fitting a uniform distribution to the sample and treat the fitted model as
# our uncertain value.
uv = UncertainValue(Uniform, some_sample)
```

You can the access the fitted model by `uv.distribution`, which on this particular random sample would give you a two-parameter model: `FittedDistribution{Uniform{Float64}}(Uniform{Float64}(a=0.001019531622161507, b=0.998951965741667))`.

You could also fit a normal distribution by running

```julia
uv = UncertainValue(Normal, some_sample)
```

Again, accessing `uv.distribution` gives you `FittedDistribution{Normal{Float64}}(Normal{Float64}(μ=0.5005522321062404, σ=0.28621018213736704))` for this particular random vector.

These fitted models would be treated just as any other theoretical distribution, discrete population or other uncertain value types. The approach is always to resample. Hence,
when doing

```julia
somesample1 = rand(10000)
somesample2 = rand(10000)
# Fit the data
uv1 = UncertainValue(Normal, some_sample)
uv2 = UncertainValue(Normal, some_sample)

# Resample uv1 and uv2 independetly, then compute the correlation on the draws
cor(uv1, uv2)
```

fitting is not done as part of the `cor` computation. This is something the user would have
to do themselves before computing the correlation, if they desire to represent their data by a fitted model.

Although not implemented yet, Julia’s multiple dispatch may optionally be used to compute certain quantities in more efficient ways when there exists analytical expressions for the uncertain value type(s) involved. If some statistic `f` is possible to compute analytically, it is of course possible to define `f(type::T, x::FittedDistribution{Normal{Float64}}, y::FittedDistribution{Normal{Float64}}) where {T <: ExactComputationType}`. This way, you could provide an extra argument indicating that an exact computation method should be used. If some statistic requires fitting a bivariate normal distribution to `x`, `y`, then this would be the way to do it.

Just calling `f(x, y)`, however, will always just resample `x` and `y` independently, then apply `f`.

Does that answer your question, @ahwillia?

### Possible future functionality

For now, only single model fits are possible. Distributions are as of now fitted using the machinery in `Distributions.jl`, using maximum likehood estimation. More sophisticated model fitting schemes are possible to implement, which I guess is where some Bayesian statistics may come in handy (e.g. the Black Box Variational Inference algorithm (Ranganath et al. 2013) you mention). I am very excited about that idea.

I imagine having an `UncertainValue` constructor with the signature `UncertainValue(::SomeBayesianScheme, x::Vector{T}) where {T <: Real}`. Instead of returning a single fitted model, this constructor would perform model selection and represent the data using a class of models that have been selected based on whatever sampling scheme you provide.

Then, I guess that the user could do

```julia
# Pretend these are measured samples with an unknown distribution
somesample1 = rand(10000)
somesample2 = rand(10000)

# Set up some Bayesian analysis
model_scheme = SomeBayesianScheme()

# Fit models to each of the measured samples using the Bayesian scheme
uval1 = UncertainValue(model_scheme, somesample1)
uval2 = UncertainValue(model_scheme, somesample2)
```

This constructor could return an instance of a new type, e.g. `SomeBayesianEstimate`, that contains the constructed model. The user could then do

```julia
uval1.model
uval2.model
```

to access the constructed models. Resampling those models is just then a matter of defining `resample(uval::SomeBayesianEstimate)`, which would sample from the class of fitted models according to some criteria.

This would be an incredibly powerful tool, because you could start mixing different model fitting methods. For example:

```
# Pretend these are measured samples with an unknown distribution
somesample1 = rand(10000)
somesample2 = rand(10000)

uval1 = UncertainValue(Normal, somesample1)
uval2 = UncertainValue(FancyBayesianScheme(), somesample2)

# Compute some statistic `f` between the two samples assuming that the first one
# is well-approximated by a Normal distribution, while the second is represented
# by a class of models derived using a Bayesian method. This would work seamlessly,
# because `resample` would be called independently on `uval1` and `uval2`.
f(uval1, uval2)
```

Thus, the resampling machinery already implemented lends itself naturally to the Bayesian approach. Although this approach is not implemented yet, it is definitiely an interesting potential future development.

### Bayesian model selection

I’ve had a brief look at the papers you suggested, @ahwillia, but they are pretty dense. Bayesian statistics is relatively new to me, and I imagine it would take a considerable effort to get my understanding of these methods to a level where I can implement them or write something reasonable about them.

Considering that fitted models is just one of many ways to represent uncertain data in the package, I am a bit hesitant to spend a lot of time on this for the current version of the package. Was your intention that a frequentist vs. Bayesian discussion needs to be included in the current release before publication of the paper, or just a hint about possible future evolution of the package documentation, @ahwillia?

Maybe improving the section on model fitting (in the documentation, this now lies under "Uncertain Values" -> "Representations" -> "Theoretical distributions with fitted parameters") is sufficient, mentioning that other approaches is possible to implement? Any thoughts?

## Combining quantitative and qualitative constraints

In summary, by using the resampling approach, combining quantitative constraints (i.e. theoretical distributions) and qualitative constraints (expert opinion) can be done seamlessly. This gives the user a tool to incorporate everything they know about their data, quantitatively and qualitatively, in their analysis. In the preprocessing stage, this can involve any kind of model fitting (if implemented) or expert judgement. You might even think about a scenario where you apply two separate Bayesian model fitting algorithms, yielding two separate classes of models, then treating these as a two-element population. During resampling, you can then draw values from the space of both models, but weight the probability of drawing from the models of a particular algorithm according to how much you trust the algorithm.

This approach is common in the field of paleoclimate science, where different proxy records (with different types of uncertainties and different magnitudes of uncertainties) representing
the same physical process are often mixed to create a single proxy record for some phenomenon. For example, proxy data values may be well-explained by some class of models, but
*weighting* the models to create the final, single record requires expert opinion on the proxy methods and analytical approaches.

## Questions

One question, @oliviaguest:

- I should probably update this paper to update the fact that UncertainData.jl is now officially part of the uncertainty handling machinery in CausalityTools.jl. How do I go about doing that? Just update the main git branch and re-trigger the whedon bot?
