### 2019-09-20 #1666 @kahaaga
@oliviaguest, you are absolutely right, it does apply here! My bad :)

I guess it would be nice that the release that is tagged in the paper also contains the latest version of the paper. The process for registering a new version of a package consists of the following steps:
 - I merge a new PR and trigger the JuliaRegistrator bot. This tests the package against the remaining Julia ecosystem.
- If the tests pass, an administrator in the central Julia repository needs to manually approve the new version.
- Once the new version is merged, the julia tagbot is activated. This makes a release in my github repo.

This process may take a few days, depending on how busy the maintainers of the main repo are.

So I guess the process is: make any changes/additions @ahwillia wants, or other additions/changes, to the paper and/or documentation. Then, if the paper is accepted, release new version. Finally, trigger whedon-bot after the tagbot has made a new release, so that this is what goes into the final pdf?
