### 2019-09-20 #1666 @oliviaguest
Great! 😊
So regrading what needs to be done: I believe so, yes. Post back here when you think it's all ready.👍
