### 2019-09-20 #1666 @kahaaga
Perfect, and thanks for the swift replies! This review process is new to me, so I just needed to be sure that I understand the process properly.

I'll post back once @ahwillia has responded to my comments and I've incorporated any final changes.
