### 2019-09-20 #1666 @ahwillia
Thanks for the detailed response! No need to change anything with regard to Bayesian methods. I just wanted to mention it as a different perspective (it is not clear to me one is better than the other). Also -- I apologize, my recommended papers were not chosen very well. I would start with [Murphy's textbook](https://www.cs.ubc.ca/~murphyk/MLbook/) particularly the section on Bayesian model selection.

**Re: corr and cov functions.** You are assuming that `x` and `y` are independent then, `p(x, y) = p(x) * p(y)`. If `x` and `y` are both univariate normals the correlation/covariance will always be zero, but for other distributions this may not be the case. Am I understanding correctly? If so, perhaps the assumption of independence should be stated?
