### 2019-09-21 c0be4d9
Added `resample` methods for tuples of uncertain values
