### 2019-09-21 c19c38f
Added `resample` methods for tuples of uncertain values
