### 2019-09-21 aafa282
Removed redudant `resample` methods for `UncertainDataset`
`UncertainDataset` is just a generic dataset that inherits all the behaviour defined for `AbstractUncertainValueDataset`, so we can just rely on multiple dispatch here.
