+++ src/resampling/Resampling.jl
@@ -43,7 +43,6 @@ using Reexport
     include("uncertain_dataset/resample_abstractuncertainvaluedataset_elwise.jl")
 
     # Specialized resampling for each type of dataset.
-    include("uncertain_dataset/resample_uncertaindataset.jl")
     include("uncertain_dataset/resample_uncertaindataset_index.jl")
     include("uncertain_dataset/resample_uncertaindataset_value.jl")
     include("uncertain_dataset/resample_uncertaindataset_indexvalue.jl")
