### 2019-09-21 7c93e05
Moved constructor for `CertainValue(x::T) where {T <: Number}`
