### 2019-09-21 b0621f7
Moved constructor for `CertainValue(x::T) where {T <: Number}`
