### 2019-09-21 335b5af
Removed too specific `resample` methods
These were just useful to document the behaviour for specific sampling constraints. We should rely on the documentation for the more generic type here.
