+++ src/resampling/uncertain_dataset/resample_uncertaindataset_index.jl
@@ -6,12 +6,12 @@ import ..UncertainDatasets.AbstractUncertainIndexDataset
 ##############################
 
 """
-	resample(uvd::UncertainIndexDataset)
+	resample(uvd::AbstractUncertainIndexDataset)
 
-Draw a realisation of an `UncertainIndexDataset` according to the distributions
-of the `UncertainValue`s comprising it.
+Draw a realisation of an uncertain index dataset according to the distributions
+of the uncertain values comprising it.
 """
-function resample(uvd::UncertainIndexDataset)
+function resample(uvd::DT) where {DT <: AbstractUncertainIndexDataset}
 	L = length(uvd)
 	[resample(uvd.indices[i]) for i in 1:L]
 end
@@ -20,8 +20,8 @@ end
 	resample(uv::AbstractUncertainIndexDataset, constraint::Vector{<:SamplingConstraint})
 
 Draw a realisation of an uncertain index dataset according to the uncertain values 
-comprising it, but constrain the elements in the dataset according the provided 
-sampling constraints.
+comprising it, while constraining the distributions furnishing the values in the dataset 
+according the provided sampling constraints.
 """
 function resample(uv::DT, constraint::Vector{<:SamplingConstraint}) where {
 		DT <: AbstractUncertainIndexDataset}
@@ -53,12 +53,9 @@ end
 Resample each element in `uvals` `n` times. The i-th entry in the returned 
 vector is a `n`-element vector consisting of `n` unique draws of `uvals[i]`.
 """
-function resample_elwise(uvd::UncertainIndexDataset, n::Int)
+function resample_elwise(uvd::DT, n::Int) where {DT <: AbstractUncertainIndexDataset}
     [resample(uvd[i], n) for i = 1:length(uvd)]
 end
-
-
-
 
 """
 	resample(uvd::UncertainIndexDataset, n::Int)
@@ -66,147 +63,136 @@ end
 Draw `n` realisations of an `UncertainIndexDataset` according to the
 distributions of the `UncertainValue`s comprising it.
 """
-function resample(uvd::UncertainIndexDataset, n::Int)
+function resample(uvd::DT, n::Int) where {DT <: AbstractUncertainIndexDataset}
 	L = length(uvd)
 	[[resample(uvd.indices[i]) for i in 1:L] for k in 1:n]
 end
 
+# ##########################################################################
+# # Draw realisations of the uncertain dataset according to the distributions
+# # of the uncertain values comprising it, according to sampling constraints
+# ##########################################################################
 
-"""
-	resample_elwise(uvd::UncertainIndexDataset, n::Int)
+# """
+# 	resample(uv::UncertainIndexDataset, constraint::NoConstraint) -> Vector{Float64}
 
-Resample each element in `uvals` `n` times. The i-th entry in the returned 
-vector is a `n`-element vector consisting of `n` unique draws of `uvals[i]`.
-"""
-function resample_elwise(uvd::UncertainIndexDataset, n::Int)
-    [resample(uvd[i], n) for i = 1:length(uvd)]
-end
+# Draw a realisation of an `UncertainIndexDataset` where each uncertain value
+# is truncated according to `constraint`. In the case of `NoConstraint`,
+# no trucation is performed and the whole distribution is sampled.
+# """
+# function resample(uv::DT, constraint::NoConstraint) where {
+# 		DT <: AbstractUncertainIndexDataset}
+# 	L = length(uv)
+# 	[resample(uv.indices[i], constraint) for i in 1:L]
+# end
 
+# """
+# 	resample(uv::UncertainIndexDataset, n::NoConstraint,
[…]
