+++ src/sampling_constraints/constrain_uncertainvaluedataset.jl
@@ -5,24 +5,57 @@ import ..UncertainDatasets:
 import ..UncertainDatasets:ConstrainedUncertainValueDataset
 
 """
-	constrain(udata::UncertainValueDataset, 
+	constrain(udata::Vector{<:AbstractUncertainValue}, 
+		s::SamplingConstraint) -> ConstrainedUncertainValueDataset
+
+Return a vector of uncertain value by applying the constraint `s` to each
+uncertain value in `udata`.
+"""
+function constrain(udata::Vector{<:AbstractUncertainValue}, constraint::SamplingConstraint)
+	[constrain(uval, constraint) for uval in udata]
+end
+
+"""
+	constrain(udata::Vector{<:AbstractUncertainValue}, 
+		constraints::Vector{T}) where {T<:SamplingConstraint} -> Vector{<:AbstractUncertainValue}
+
+Return a vector of uncertain values by applying a different sampling constraint to each uncertain 
+value in `udata`.
+"""
+function constrain(udata::Vector{<:AbstractUncertainValue}, 
+		constraints::Vector{T}) where {T <: SamplingConstraint}
+	
+	if length(udata) != length(constraints)
+		error("Number of sampling constraints must match length of dataset.")
+	end
+
+	return [constrain(udata[i], constraints[i]) for i in 1:length(udata)]
+end
+
+"""
+	constrain(udata::AbstractUncertainValueDataset, 
 		s::SamplingConstraint) -> ConstrainedUncertainValueDataset
 
 Return a uncertain dataset by applying the constraint `s` to each
 uncertain value in `udata`.
 """
-constrain(udata::UncertainValueDataset, constraint::SamplingConstraint) = 
+function constrain(udata::DT, constraint::SamplingConstraint) where {
+		DT <: AbstractUncertainValueDataset}
+	
 	ConstrainedUncertainValueDataset([constrain(uval, constraint) for uval in udata])
-
+end
 
 """
-	constrain(udata::UncertainValueDataset, 
+	constrain(udata::AbstractUncertainValueDataset, 
 		constraints::Vector{T}) where {T<:SamplingConstraint} -> ConstrainedUncertainValueDataset
 
 Return a uncertain dataset by applying a different sampling constraint to each uncertain 
 value in `udata`.
 """
-function constrain(udata::UncertainValueDataset, constraints::Vector{T}) where {T<:SamplingConstraint}
+function constrain(udata::DT, constraints::Vector{T}) where {
+		T<:SamplingConstraint, 
+		DT <: AbstractUncertainValueDataset}
+	
 	if length(udata) != length(constraints)
 		error("Number of sampling constraints must match length of dataset.")
 	end
@@ -33,34 +66,34 @@ function constrain(udata::UncertainValueDataset, constraints::Vector{T}) where {
 end
 
 
-"""
-	constrain(udata::ConstrainedUncertainValueDataset, 
-		constraint::SamplingConstraint) -> ConstrainedUncertainValueDataset
+# """
+# 	constrain(udata::ConstrainedUncertainValueDataset, 
+# 		constraint::SamplingConstraint) -> ConstrainedUncertainValueDataset
 
-Return a uncertain dataset by applying the `constraint` to each
-uncertain value in `udata`.
-"""
-constrain(udata::ConstrainedUncertainValueDataset, constraint::SamplingConstraint) = 
-	ConstrainedUncertainValueDataset([constrain(uval, constraint) for uval in udata])
+# Return a uncertain dataset by applying the `constraint` to each
+# uncertain value in `udata`.
+# """
+# constrain(udata::ConstrainedUncertainValueDataset, constraint::SamplingConstraint) = 
[…]
