### 2019-09-21 8bce67b
Identity constructor
