### 2019-09-21 d690c47
Add some convert methods for certain values
