### 2019-09-21 4d3563c
Docs: new requirements for `AbstractUncertainIndexDataset`
