### 2019-09-21 7fc7625
Return vector directly when resampling vector of uncertain values
