+++ src/resampling/uncertain_vectors/resample_uncertain_vectors.jl
@@ -29,3 +29,15 @@ function resample(uv::DT, constraint::SamplingConstraint, n::Int) where {
         DT <: Vector{<:AbstractUncertainValue}}
     [[resample(uv[i], constraint) for i in 1:length(uv)] for k = 1:n]
 end
+
+function resample(uv::DT, constraint::NoConstraint) where {
+    DT <: Vector{<:AbstractUncertainValue}}
+
+    return uv
+end
+
+function resample(uv::DT, constraint::NoConstraint, n::Int) where {
+    DT <: Vector{<:AbstractUncertainValue}}
+    
+    return [uv for i = 1:n]
+end
