### 2019-09-22 4e25cf0
Added missing constructors for `UncertainIndexValueDataset`
