### 2019-09-22 313bcfb
Update docs for `cor` and `cov` methods for uncertain value datasets
