### 2019-09-22 0b10b69
new `resample` methods for uncertain population
