### 2019-09-22 d4fd0ef
Update docs for UncertainIndexValueDataset
