### 2019-09-22 70e3e00
Some restructuring
