+++ src/uncertain_values/abstract_types/AbstractPopulation.jl
@@ -1,0 +1,8 @@
+"""
+    AbstractPopulation
+
+An abstract type for population-based uncertain values.
+"""
+abstract type AbstractPopulation <: AbstractUncertainValue end 
+
+export AbstractPopulation
