+++ src/uncertain_values/abstract_types/AbstractUncertainScalarKDE.jl
@@ -1,0 +1,2 @@
+
+abstract type AbstractUncertainScalarKDE{T} <: AbstractEmpiricalValue end
