### 2019-09-22 3dc4b0b
Some restructuring
