+++ src/uncertain_values/UncertainValues.jl
@@ -13,16 +13,28 @@ using Reexport
         Weights, weights,
         AnalyticWeights, aweights
 
-  
-    # Abstract types and common methods
-    include("AbstractUncertainValue.jl")
-    include("AbstractEmpirical.jl")
+    ##########################################
+    # Abstract types
+    ##########################################
+    include("abstract_types/AbstractUncertainValue.jl")
 
+    include("abstract_types/AbstractPopulation.jl")
+    include("abstract_types/AbstractScalarPopulation.jl")
+    include("abstract_types/AbstractEmpirical.jl")
+    include("abstract_types/AbstractUncertainScalarKDE.jl")
+
+
+    ##########################################
     # Distributions
+    ##########################################
     include("distributions/assign_dist.jl")
 
-    # Fitted distribution type 
+    ##########################################
+    # Fitted distribution type
+    ##########################################
     include("distributions/fitted_distribution.jl")
+
+
 
     ##########################################
     # Composite uncertain scalar types
@@ -37,7 +49,7 @@ using Reexport
     include("UncertainScalarsKDE.jl")
 
     # Populations with weighted probabilities
-    include("Population.jl")
+    include("UncertainScalarPopulation.jl")
     
     # Certain values (i.e. values without uncertainty)
     include("CertainValue.jl")
@@ -53,6 +65,11 @@ using Reexport
     # Define common constructor, so a similar syntax may be used to construct
     # all types of uncertain values.
     include("UncertainValue.jl")
+
+    ##########################################
+    # Comparisons
+    ##########################################
+    include("operations/comparisons.jl")
 
     # Conversions
     include("convert.jl")
