### 2019-09-22 7d3702c
Add iteration to UncertainScalarPopulation
