### 2019-09-22 6b77297
Add iteration to UncertainScalarPopulation
