### 2019-09-24 #1666 @kahaaga
## Bayesian stuff

Thanks for the reference to Murphy's textbook. Although model fitting is not the main focus of this package, bayesian model selection would be an interesting feature to have when the user wants to model their empirical distributions (beyond what is already possible in the package). I'll check the book and the papers you cited in more detail .

## About `cor` and `cov`

Actually, neither `cor` nor `cov` are implemented for single uncertain values, only for uncertain datasets . The function signature for the correlation is now`cor(d1::AbstractUncertainValueDataset, d2::AbstractUncertainValueDataset, n::Int)`.

That is, they are computed on draws from collections of different values, where each value in the collection can be certain (e.g. a scalar) or uncertain (theoretical normal distribution, a fitted distribution, a population, etc).

This would be the equivalent of having, say, two time series with uncertainties in the measured values. Maybe the magnitudes of the uncertainties fluctuate through time, so at each instant of time there is a different distribution representing the value.

## Compute a statistic between two *single* uncertain values
If I wanted point estimates on a statistic `f`, I could just define a method `f(uval1::AbstractUncertainValue, uval2::AbstractUncertainValue, n::Int)` that computes the correlation between two single uncertain values `uval1` and `uval2` by resampling them independently `n` time, then computing `f` on the draws. I [describe how to do this](https://kahaaga.github.io/UncertainData.jl/dev/implementing_algorithms_for_uncertaindata/) in the documentation.

For example, define `cor(uval1::AbstractUncertainValue, uval2::AbstractUncertainValue, n::Int)` as

```julia
import Statistics: cor

function cor(uval1::AbstractUncertainValue, uval2::AbstractUncertainValue, n::Int)
    cor(resample(uval1, n), resample(uval2, n))
end
```

Which we can use like this:

```julia
# Point-estimate of correlation between a gamma distribution and a
# three-member weighted population with relative probabilities 0.3, 0.3 and 0.4
cor(UncertainValue(Gamma, 1, 2), UncertainValue([1.1, 2.2, 3.3], [0.3, 0.3, 0.4]), 100)
> 0.007348007353157689

# Point-estimate of the correlation between two normal distributions
cor(UncertainValue(Normal, 0, 1), UncertainValue(Normal, 4.2, 0.5), 10000)
> -0.002492518462927909

# Define a mixture model of a bunch of different distributions, and sample this
# to simulate some measured value with complicated uncertainty
M1 = MixtureModel([Beta(0.5, 1), Normal(-3, 0.5), Normal(2, 1), Normal(5, 2), Gamma(1, 2)])
empirical_sample = rand(M1, rand(5000:10000))

# Point-estimate of the correlation between this empirical sample and a
# beta distribution (calling UncertainValue(empirical_sample) triggers a KDE estimate
# to the distribution.
cor(UncertainValue(empirical_sample), UncertainValue(Beta, 2, 5), 10000)
> 0.013776522799511042
```

Hence, I get a point estimate for my statistic that is the result of sampling each value `n=10000` times, then computing the correlation between those two `10000`-member distributions.

Here, I assume nothing other than that `uval1` and `uval2` must be sampled independently to not induce any correlation that is not present.

This is different from computing the correlation between two uncertain *datasets*, which I describe below.

## Compute a statistic between two uncertain *datasets*

### Example

```julia
using UncertainData, Distributions

# Simulate some mean data values measured for two phenomena (whatever they are)
values_var1 = sin.(0:0.1:10)
values_var2 = cos.(0:0.1:10)
```
If we wanted, we could now directly compute the correlation between the two time series
by calling `cor(values_var1, values_var2)`. This would yield a single estimate of the correlation statistic:

```julia
cor(values_var1, values_var2)
> 0.0548362735517299
```

But instead, let's imagine that we also had access to information about uncertainties on the measure data values.

```julia
# Assume the data were measured by a device with normally distributed measurement uncertainties with fluctuating standard deviations in the following range
σ_range = (0.1, 0.7)
```

Now, how do we compute the correlation between the two series, when each of them have
uncertainties associated with each point in the time series? The answer (in this package) is
to resample.

```julia
# Instead of working with just the means, let's instead work with each point as
# a normal distribution.
uncertain_vals1 = [UncertainValue(Normal, val, rand(Uniform(σ_range...))) for val in values_var1]
uncertain_vals2 = [UncertainValue(Normal, val, rand(Uniform(σ_range...))) for val in values_var2]

# Convert to `UncertainValueDataset`s, so that we can call `cor`
uncertain_dataset1 = UncertainValueDataset(uncertain_vals1)
uncertain_dataset2 = UncertainValueDataset(uncertain_vals2)
```

Plotting the data:

Now, draw one realisation of `uncertain_dataset1` in an element-wise manner (one draw from each distribution furnishing the dataset). Next, draw an independent realisation of `uncertain_dataset2` the same way. Finally, compute the correlation between these two realisations. The function signature for the correlation is now`cor(d1::AbstractUncertainValueDataset, d2::AbstractUncertainValueDataset, n::Int)`, so to estimate the correlation for a single draw, you´d do

```julia
cor(uncertain_dataset1, uncertain_dataset2, 1)
```

which would give you a single estimate for the correlation. Repeating that many times, we get a distribution of estimates for the correlation:

```julia
cor(uncertain_dataset1, uncertain_dataset2, 10000)
```

Plotting the histogram of the distribution of correlation estimates, we get a mean correlation estimate close to the estimate for the time series without uncertainties.

Thus, there's nothing special about `cor` or `cov`. Using this resampling strategy, any other quantity or statistic can be computed in a similar manner. This is exactly the same approach I use in [`CausalityTools`](https://github.com/kahaaga/CausalityTools.jl) to obtain confidence intervals on the various statistical estimators of dynamical coupling, which are alternative ways of analyzing time series like this.

Of course, the *interpretation* of the results is something else entirely, and here you'd have think about underlying assumptions. One can always use this resampling approach to compute *anything*. How meaningful it is to do so for a given statistic, such as the correlation, is a different question. It is the user's responsibility to know the limitations of the statistical estimators they apply, what they mean, and how applicable the method is to their data.

The advantage of this approach is that it separates the analysis into distinct stages for the user:

1. Define the empirical distributions, weighted populations, fitted distributions or theoretical distributions representing the possible values in the dataset. Here, you can incorporate most of the prior information you have about your data.

2. Implement the statistic you're interested in for vector-valued inputs.

3. Make a wrapper around the function for the statistic that accepts uncertain datasets as inputs.

4. Repeat the computation for multiple realisations of your dataset. Resampling many times, we get a distribution over the statistic, from which we can obtain confidence intervals on the statistic.

In the resampling stage, you can also input other prior information, such as "I only want to sample the indices of my dataset such that they are strictly increasing". The statistic is then computed over the draws, and we get a estimated distribution over the value of the statistic. This distribution then encompasses all the prior information you had on uncertainties and on sampling constraints that you applied during the resampling process.

Does this answer your question?

I will implement the `cor(x::AbstractUncertainValue, y::AbstractUncertainValue, n::Int)` method as described above as part of the next release (so that it gets into the version that goes with the paper).

Although the information is already present in the documentation, I think it is a good idea to make some tutorials that describes this more thoroughly, like in this answer. I'll create an issue in the github repository that tracks this.
