### 2019-09-24 kahaaga/UncertainData.jl#92 @kahaaga
Tutorial explaining the difference between point-estimates for a statistic and estimates of the statistic for uncertain datasets 
Ref. discussion in https://github.com/openjournals/joss-reviews/issues/1666
