### 2019-09-24 kahaaga/UncertainData.jl#93 @kahaaga
`cor` and `cov` point-estimates
Ref discussion in https://github.com/openjournals/joss-reviews/issues/1666.

Implement the following methods to compute point estimates between two single uncertain values:

- `cor(x::AbstractUncertainValue, y::AbstractUncertainValue, n::Int)`
- `cov(x::AbstractUncertainValue, y::AbstractUncertainValue, n::Int)`
