### 2019-09-24 kahaaga/UncertainData.jl#94 @kahaaga
Improve docs for `cor` and `cov` for uncertain datasets
Ref discussion in https://github.com/openjournals/joss-reviews/issues/1666.

For the following methods, the documentation can be clearer on how the resampling is done, how the statistics are computed, and what is returned:

- `cor(x::AbstractUncertainValueDataset, y::AbstractUncertainValueDataset, n::Int)`
- `cov(x::AbstractUncertainValueDataset, y::AbstractUncertainValueDataset, n::Int)`
