### 2019-09-24 #1666 @ahwillia
Interesting -- thanks for putting so much effort into this. I like the goal of breaking it down into 4 steps for the user.

> Here, I assume nothing other than that uval1 and uval2 must be sampled independently

^ This was all I was after. I think if you don't make some sort of assumption like this, it can be a hard problem.

For example, you could think of drawing `n` observations from a multivariate Gaussian distribution. Say this results in a data matrix with shape `n x m`.

Then, one could fit the mean and covariance matrix to those data to estimate the underlying distribution.

Now, when you call `resample`, I suppose it would return entire rows (length-`m`) vectors. By doing this it would capture correlations amongst the `m` variables, which wouldn't be possible if you resampled each of the `m` columns independently. In this case the columns **are not independent of each other** due to the correlations induced by the covariance matrix.

> Of course, the interpretation of the results is something else entirely, and here you'd have think about underlying assumptions. One can always use this resampling approach to compute anything. How meaningful it is to do so for a given statistic, such as the correlation, is a different question. It is the user's responsibility to know the limitations of the statistical estimators they apply, what they mean, and how applicable the method is to their data.

I agree with this, assuming that `resample` draws an actual sample from the true underlying distribution of the data. So all I request is that the assumptions are stated in the tutorials.

For example, the fact that noise is applied elementwise in the time series example is an assumption. Alternatively, one could fit a Gaussian process model to the data. Resamples from this model would be very different than resamples from the i.i.d. noise model you suggest.

Thanks again for putting time into making detailed tutorials for your package. Again, I am signed off for this to be accepted.
