### 2019-09-25 acf553d
Not all subtypes of `AbstractUncertainValue` has the `distribution` field
