### 2019-09-25 d400090
Not all subtypes of `AbstractUncertainValue` has the `distribution` field
