+++ src/uncertain_values/UncertainScalarPopulation.jl
@@ -3,32 +3,50 @@ import IntervalArithmetic: interval
 import Distributions 
 import StatsBase
 
-"""
-    UncertainScalarPopulation
-
-An uncertain value represented by a population for which the probabilities of the values 
-are dictated by a set of weights. The weights are normalized by default.
-
-
-## Examples 
-
-The two following ways of constructing populations are equivalent. 
-
-```julia 
-values = rand(1:50, 100)
-weights = rand(100)
-population = UncertainValue(values, weights)
-```
-
-```julia 
-values = rand(1:50, 100)
-weights = rand(100)
-population = UncertainScalarPopulation(values, weights)
-```
+const POTENTIAL_UVAL_TYPES = Union{T1, T2} where {T1<:Number, T2} where  T2 <: AbstractUncertainValue
 
 """
-struct UncertainScalarPopulation{T, PW <: StatsBase.AbstractWeights} <: AbstractScalarPopulation
-    values::Vector{<:T}
+    UncertainScalarPopulation(values, probs)
+    UncertainScalarPopulation(values, probs::Vector{Number})
+    UncertainScalarPopulation(values, probs::Statsbase.AbstractWeights)
+
+
+An `UncertainScalarPopulation`, which consists of some population members (`values`)
+and some weights (`probs`) that indicate the relative importance of the 
+population members (for example during resampling). 
+
+There are different constructors for different types of `values`:
+
+- If `values` contains only scalar numeric values, then the `values` field 
+    will be of type `Vector{Number}`.
+- If `values` contains one or more uncertain values, then the `values` field 
+    will be of type `Vector{AbstractUncertainValue}`
+
+## Example 
+
+```julia
+
+# Uncertain population consisting of CertainValues (scalars get promoted to 
+# CertainValue), theoretical distributions and KDE distributions
+pop1 = UncertainScalarPopulation(
+    [3.0, UncertainValue(Normal, 0, 1), UncertainValue(Gamma, 2, 3), 
+    UncertainValue(Uniform, rand(1000))], [0.5, 0.5, 0.5, 0.5])
+
+# Uncertain population consisting of scalar values
+pop2 = UncertainScalarPopulation([1, 2, 3], rand(3))
+pop3 = UncertainScalarPopulation([1, 2, 3], Weights(rand(3)))
+
+# Uncertain population consisting of uncertain populations
+pop4 = UncertainScalarPopulation([pop1, pop2], [0.1, 0.5])
+
+# Uncertain population consisting of uncertain populations, a scalar and 
+# a normal distribution. Assign random weights.
+vals = [pop1, pop2, 2, UncertainValue(Normal, 0.3, 0.014)]
+pop5 = UncertainScalarPopulation(vals, Weights(rand(4)))
+```
+"""
+struct UncertainScalarPopulation{T, PW <: StatsBase.AbstractWeights} <: AbstractScalarPopulation{T, PW}
+    values::Vector{T}
     probs::PW
 end
 
@@ -36,9 +54,14 @@ end
     UncertainScalarPopulation(values::Vector, probabilities::Vector{Float64})
 
 Construct a population from a vector of values and a vector of probabilities associated 
-to those values. The weights are normalized by default.
-"""
-function UncertainScalarPopulation(values::Vector, probabilities::Vector) where {T2}
+to those values."""
+function UncertainScalarPopulation(values::Vector{T1}, probabilities::Vector{T2}) where {T1 <: Number, T2}
+    if length(values) != length(probabilities)
+        throw(ArgumentError("Lengths of values and probability vectors do not match."))
+    end
+    UncertainScalarPopulation(values, StatsBase.weights(probabilities))
+end
[…]
