### 2019-09-25 05ed739
Added `resample` method for evaluating point-estimates for single uncertain values
