### 2019-09-25 7b07bbc
Added `resample` method for evaluating point-estimates for single uncertain values
