+++ src/statistics/statsbase/uncertain_datasets/core_stats_uncertaindatasets.jl
@@ -82,6 +82,19 @@ Statistics.cor(d1::DT, d2::DT, n::Int; kwargs...) where {DT <: AbstractUncertain
     [cor(resample(d1), resample(d2), kwargs...) for i = 1:n]
 
 """
+    cor(d1::AbstractUncertainValueDataset, d2::AbstractUncertainValueDataset; kwargs...)
+
+Obtain a single estimate of the Pearson correlation between two uncertain datasets `d1` and `d2`.
+This is done by resampling both datasets independently, that is: First draw a realisation of 
+`d1` according to the distributions furnishing its uncertain values. Then, draw a realisation 
+`d2` according to its furnishing distributions. These those two draws/realisations are now 
+both vectors of length `L = length(d1) = length(d2)`. Finally, compute the correlation between 
+those draws. This yields a single estimates of the correlation between `d1` and `d2`.
+"""
+Statistics.cor(d1::DT, d2::DT, kwargs...) where {DT <: AbstractUncertainValueDataset} =
+    cor(resample(d1), resample(d2), kwargs...)
+
+"""
     cov(d1::AbstractUncertainValueDataset, d2::AbstractUncertainValueDataset, n::Int; kwargs...)
 
 Obtain a distribution for the covariance between two uncertain datasets `d1` and `d2`.
@@ -98,8 +111,21 @@ The procedure is as follows.
     This yields `n` estimates of the covariance between `d1` and `d2`, which is returned as 
     a vector.
 """
-Statistics.cov(d1::AbstractUncertainValueDataset, d2::AbstractUncertainValueDataset, n::Int; kwargs...) =
+Statistics.cov(d1::DT, d2::DT, n::Int; kwargs...) where DT <: AbstractUncertainValueDataset =
     [cov(resample(d1), resample(d2), kwargs...) for i = 1:n]
 
+
+"""
+    cov(d1::AbstractUncertainValueDataset, d2::AbstractUncertainValueDataset; kwargs...)
+
+Obtain a single estimate for the covariance between two uncertain datasets `d1` and `d2`.
+This is done by resampling both datasets independently, that is: First draw a realisation of 
+`d1` according to the distributions furnishing its uncertain values. Then, draw a realisation 
+`d2` according to its furnishing distributions. These those two draws/realisations are now 
+both vectors of length `L = length(d1) = length(d2)`. Finally, compute the covariance between 
+those draws. This yields a single estimates of the covariance between `d1` and `d2`.
+"""
+Statistics.cov(d1::DT, d2::DT, kwargs...) where {DT <: AbstractUncertainValueDataset} =
+    cov(resample(d1), resample(d2), kwargs...)
 
 export cor, cov, var, std, quantile, mean, median
