### 2019-09-25 97d34f7
Added some methods for `CertainValue`
