### 2019-09-25 7bcba16
Deprecate core stats methods
