### 2019-09-25 ab81c54
Rework `resample` methods using the new `resample` method for point-estimates
