### 2019-09-25 c67de56
Add generic `resample` function without arguments
