### 2019-09-25 223893e
Add point estimators to changelog
