### 2019-09-26 d4dc13c
Extend more pointwise and pairwise statistical estimators
