### 2019-09-26 fc3363f
Extend more pointwise and pairwise statistical estimators
