+++ src/statistics/statsbase/uncertain_values/core_stats_point_estimates.jl
@@ -1,0 +1,174 @@
+import StatsBase
+
+"""
+	mean(uv::AbstractUncertainValue, n::Int)
+
+Compute the mean of an uncertain value over an `n`-draw sample of it.
+"""
+StatsBase.mean(x::AbstractUncertainValue, n::Int) = resample(StatsBase.mean, x, n)
+
+"""
+	mode(uv::AbstractUncertainValue, n::Int)
+
+Compute the mode of an uncertain value over an `n`-draw sample of it.
+"""
+StatsBase.mode(x::AbstractUncertainValue, n::Int) = resample(StatsBase.mode, x, n)
+
+"""
+	quantile(uv::AbstractUncertainValue, q, n::Int)
+
+Compute the quantile(s) `q` of an uncertain value over an `n`-draw sample of it.
+"""
+StatsBase.quantile(x::AbstractUncertainValue, q, n::Int) = resample(StatsBase.quantile, x, n, q)
+
+"""
+	iqr(uv::AbstractUncertainValue, n::Int)
+
+Compute the interquartile range (IQR), i.e. the 75th percentile minus the 25th percentile,
+over an `n`-draw sample of an uncertain value.
+"""
+StatsBase.iqr(x::AbstractUncertainValue, n::Int) = resample(StatsBase.iqr, x, n, q)
+
+
+"""
+	median(uv::AbstractUncertainValue, n::Int)
+
+Compute the median of an uncertain value over an `n`-draw sample of it.
+"""
+StatsBase.median(x::AbstractUncertainValue, n::Int) = resample(StatsBase.median, x, n)
+
+"""
+	middle(uv::AbstractUncertainValue, n::Int)
+
+Compute the middle of an uncertain value over an `n`-draw sample of it.
+"""
+StatsBase.middle(x::AbstractUncertainValue, n::Int) = resample(StatsBase.middle, x, n)
+
+"""
+	std(uv::AbstractUncertainValue, n::Int)
+
+Compute the standard deviation of an uncertain value over an `n`-draw sample of it.
+"""
+StatsBase.std(x::AbstractUncertainValue, n::Int) = resample(StatsBase.std, x, n)
+
+"""
+	variance(uv::AbstractUncertainValue, n::Int)
+
+Compute the variance of an uncertain value over an `n`-draw sample of it.
+"""
+StatsBase.var(x::AbstractUncertainValue, n::Int) = resample(StatsBase.var, x, n)
+
+"""
+	genmean(uv::AbstractUncertainValue, p, n::Int)
+
+Compute the generalized/power mean with exponent `p` of an uncertain value over an 
+`n`-draw sample of it.
+"""
+StatsBase.genmean(x::AbstractUncertainValue, p, n::Int) = resample(StatsBase.genmean, x, n, p)
+
+"""
+	genvar(uv::AbstractUncertainValue, n::Int)
+
+Compute the generalized sample variance of an uncertain value over an 
+`n`-draw sample of it.
+"""
+StatsBase.genvar(x::AbstractUncertainValue, n::Int) = resample(StatsBase.genvar, x, n)
+
+"""
+	harmmean(uv::AbstractUncertainValue, n::Int)
+
+Compute the harmonic mean of an uncertain value over an `n`-draw sample of it.
+"""
+StatsBase.harmmean(x::AbstractUncertainValue, n::Int) = resample(StatsBase.harmmean, x, n)
+
+"""
+	geomean(uv::AbstractUncertainValue, n::Int)
+
+Compute the geometric mean of an uncertain value over an `n`-draw sample of it.
+"""
+StatsBase.geomean(x::AbstractUncertainValue, n::Int) = resample(StatsBase.geomean, x, n)
+
+"""
+	kurtosis(uv::AbstractUncertainValue, n::Int)
+
[…]
