+++ src/statistics/statsbase/uncertain_values/core_stats_pair_estimates.jl
@@ -1,0 +1,193 @@
+"""
+    countne(x::AbstractUncertainValue, y::AbstractUncertainValue, n::Int)
+
+Count the number of indices at which the elements of two independent length-`n`
+draws of `x` and for `y` are not equal. 
+"""
+StatsBase.countne(x::AbstractUncertainValue, y::AbstractUncertainValue, n::Int) = 
+    resample(StatsBase.countne, x, y, n)
+
+"""
+    counteq(x::AbstractUncertainValue, y::AbstractUncertainValue, n::Int)
+
+Count the number of indices at which the elements of two independent length-`n`
+draws of `x` and for `y` are equal. 
+"""
+StatsBase.counteq(x::AbstractUncertainValue, y::AbstractUncertainValue, n::Int) = 
+    resample(StatsBase.counteq, x, y, n)
+
+"""
+    corkendall(x::AbstractUncertainValue, y::AbstractUncertainValue, n::Int)
+
+Compute Kendalls's rank correlation coefficient between two uncertain values by 
+independently drawing `n` samples from `x` and `n` samples from `y`, then computing 
+Kendalls's rank correlation coefficient between those length-`n` draws.
+"""
+StatsBase.corkendall(x::AbstractUncertainValue, y::AbstractUncertainValue, n::Int) = 
+    resample(StatsBase.corspearman, x, y, n)
+
+"""
+    corspearman(x::AbstractUncertainValue, y::AbstractUncertainValue, n::Int)
+
+Compute Spearman's rank correlation coefficient between two uncertain values by 
+independently drawing `n` samples from `x` and `n` samples from `y`, then computing 
+the Spearman's rank correlation coefficient between those length-`n` draws.
+"""
+StatsBase.corspearman(x::AbstractUncertainValue, y::AbstractUncertainValue, n::Int) = 
+    resample(StatsBase.corspearman, x, y, n)
+
+"""
+    cor(x::AbstractUncertainValue, y::AbstractUncertainValue, n::Int)
+
+Compute the Pearson correlation between two uncertain values by independently 
+drawing `n` samples from `x` and `n` samples from `y`, then computing 
+the Pearson correlation between those length-`n` draws.
+"""
+StatsBase.cor(x::AbstractUncertainValue, y::AbstractUncertainValue, n::Int) = 
+    resample(StatsBase.cor, x, y, n)
+
+"""
+    cov(x::AbstractUncertainValue, y::AbstractUncertainValue, n::Int; 
+        corrected::Bool = true)
+
+Compute the covariance between two uncertain values by independently 
+drawing `n` samples from `x` and `n` samples from `y` , then computing 
+the covariance between those length-`n` draws.
+"""
+StatsBase.cov(x::AbstractUncertainValue, y::AbstractUncertainValue, n::Int; corrected::Bool = true) = 
+    resample(StatsBase.cov, x, y, n; corrected = corrected)
+
+"""
+    crosscor(x::AbstractUncertainValue, y::AbstractUncertainValue, [lags], 
+        n::Int; demean = true)
+
+Compute the cross correlation between two uncertain values by independently 
+drawing `n` samples from `x` and `n` samples from `y`, `x_draw` and `y_draw`, then computing 
+the cross correlation between those length-`n` draws. `demean` specifies whether the respective 
+means of the `x_draw` and `y_draw` should be subtracted from them before computing 
+their cross correlation.
+
+When left unspecified, the `lags` used are `-min(n-1, 10*log10(n))` to `min(n, 10*log10(n))`.
+
[…]
