### 2019-09-26 10e387b
Test UncertainScalarPopulation initialisation
