### 2019-09-26 46a22be
Test UncertainScalarPopulation initialisation
