### 2019-09-26 1f5594e
Improve and restructure resample tests
