+++ test/resampling/test_resampling_uncertain_values.jl
@@ -1,0 +1,8 @@
+
+#############################
+# Resampling uncertain values
+#############################
+include("uncertain_values/test_resampling_certain_value.jl")
+include("uncertain_values/test_resampling_uncertainvalues.jl")
+include("uncertain_values/test_resampling_uncertainvalues_kde.jl")
+include("uncertain_values/test_resampling_UncertainScalarPopulation.jl")
