+++ test/resampling/uncertain_datasets/test_resampling_datasets_uncertainindexvaluedataset.jl
@@ -1,0 +1,7 @@
+UI = UncertainValueDataset(uidxs)
+UV = UncertainValueDataset(uvals)
+UIV = UncertainIndexValueDataset(UI, UV)
+
+n = 3
+@test resample(UIV) isa Tuple{Vector}
+@test resample(UIV, n) isa Vector{Tuple{Vector}}
