+++ test/resampling/uncertain_datasets/test_resampling_datasets_uncertainindexdataset.jl
@@ -1,0 +1,6 @@
+
+UI = UncertainIndexDataset(uidxs)
+
+n = 3
+@test resample(UI) isa Vector{<:Real}
+@test resample(UI, n) isa Vector{Vector{T}} where T <: Real
