### 2019-09-26 160bf10
Improve and restructure resample tests
