+++ test/resampling/uncertain_datasets/test_resampling_datasets_uncertainindexdataset_apply_funcs.jl
@@ -1,0 +1,12 @@
+##########################################
+# Some uncertain datasets
+##########################################
+UI = UncertainIndexDataset(uidxs)
+
+##########################################
+# Apply functions to datasets `n` times
+##########################################
+n = 3
+@test resample(median, UI, n) isa Vector{T} where T <: Real
+@test resample(cor, UI, UI, n) isa Vector{T} where T <: Real
+
