+++ test/resampling/uncertain_datasets/test_resampling_datasets_uncertainvaluedataset.jl
@@ -1,0 +1,5 @@
+UV = UncertainValueDataset(uvals)
+
+n = 3
+@test resample(UV) isa Vector{T} where T <: Real
+@test resample(UV, n) isa Vector{Vector{T}} where T <: Real
