+++ test/resampling/uncertain_vectors/test_resampling_vectors.jl
@@ -1,51 +1,15 @@
-import KernelDensity.UnivariateKDE
-o1 = UncertainValue(Normal, 0, 0.5)
-o2 = UncertainValue(Normal, 2.0, 0.1)
-o3 = UncertainValue(Uniform, 0, 4)
-o4 = UncertainValue(Uniform, rand(100))
-o5 = UncertainValue(Beta, 4, 5)
-o6 = UncertainValue(Gamma, 4, 5)
-o7 = UncertainValue(Frechet, 1, 2)
-o8 = UncertainValue(BetaPrime, 1, 2)
-o9 = UncertainValue(BetaBinomial, 10, 3, 2)
-o10 = UncertainValue(Binomial, 10, 0.3)
-o11 = UncertainValue(rand(100), rand(100))
-o12 = UncertainValue(2)
-o13 = UncertainValue(2.3)
-o14 = UncertainValue([2, 3, 4], [0.3, 0.4, 0.3])
-o15 = UncertainValue([rand() for i = 1:100])
-
-
-uvals1 = [o1, o2, o3, o4, o5, o6, o7, o8, o9, o11, o12, o13, o14, o1, o2, o3, o15, o4, o5, 
-    o6, o7, o8, o9, o11, o12, o13, o14, o1, o2, o3, o4, o5, o6, o7, o8, o9, o11, o12, o13, 
-    o14, o1, o2, o3, o4, o5, o6, o7, o8, o9, o11, o12, o13, o14]
-
 
 # resample(uvals::Vector{AbstractUncertainValue})
-@test resample(uvals1) isa Vector
-@test length(resample(uvals1)) == length(uvals1)
+@test resample(uvals) isa Vector
+@test length(resample(uvals)) == length(uvals)
 
 # resample(uvals::Vector{AbstractUncertainValue}, n::Int) 
 n = 2
-@test resample(uvals1, n) isa Vector{Vector{T}} where T <: Real
-@test length(resample(uvals1, n)) == n
+@test resample(uvals, n) isa Vector{Vector{T}} where T <: Real
+@test length(resample(uvals, n)) == n
 
 # resample_elwise(uvals::Vector{AbstractUncertainValue}, n::Int) 
 n = 2
-@test resample_elwise(uvals1, n) isa Vector{Vector}
-@test length(resample_elwise(uvals1, n)) == length(uvals1)
-@test length(resample_elwise(uvals1, n)[1]) == n
-
-uvals_x = [UncertainValue(Normal, rand(Normal(0, 5)), abs(rand(Normal(0, 3)))) for i = 1:100]
-
-constraints = [[TruncateStd(0.3) for x in 1:50]; [TruncateQuantiles(0.3, 0.7) for x in 1:50]];
-
-# Draw one realisation
-@test resample(uvals_x) isa Vector{<:Real}
-@test resample(uvals_x, constraints[1]) isa Vector{<:Real}
-@test resample(uvals_x, constraints) isa Vector{<:Real}
-
-# Draw multiple realisations
-@test resample(uvals_x, 5) isa Vector{Vector{T}} where T<:Real
-@test resample(uvals_x, constraints, 5) isa Vector{Vector{T}} where T<:Real
-@test resample(uvals_x, constraints[1], 5) isa Vector{Vector{T}} where T<:Real
+@test resample_elwise(uvals, n) isa Vector{Vector}
+@test length(resample_elwise(uvals, n)) == length(uvals)
+@test length(resample_elwise(uvals, n)[1]) == n

[…]
