+++ src/resampling/uncertain_values/resample_uncertainvalues_vector_withconstraints.jl
@@ -1,6 +1,6 @@
 
 """
-resample(uvals::Vector{AbstractUncertainValue}, c::SamplingConstrant)
+    resample(uvals::Vector{AbstractUncertainValue}, c::SamplingConstrant)
 
 Treat `uvals` as a dataset and resample each value of `uvals` once,
 Returns an `length(uvals)`-element vector.
@@ -9,18 +9,18 @@ resample(uvals::Vector{AbstractUncertainValue}, c::SamplingConstrant) = resample
 
 
 """
-resample(uvals::Vector{AbstractUncertainValue}, n::Int) 
+    resample(uvals::Vector{AbstractUncertainValue}, n::Int) 
 
 Treat `uvals` as a dataset and resample it `n` times. 
 Returns `n` resampled draws of `uvals`, each being a `length(uvals)`-element vector. 
 For each returned vector, the i-th element is a unique draw of `uvals[i]`. 
 """
 function resample(uvals::Vector{AbstractUncertainValue}, n::Int) 
-[resample.(uvals) for i = 1:n]
+    [resample.(uvals) for i = 1:n]
 end
 
 """
-resample_elwise(uvals::Vector{AbstractUncertainValue}, n::Int) 
+    resample_elwise(uvals::Vector{AbstractUncertainValue}, n::Int) 
 
 Resample each element in `uvals` `n` times. The i-th entry in the returned 
 vector is a `n`-element vector consisting of `n` unique draws of `uvals[i]`.
