### 2019-09-26 3db129d
Add more `resample` methods that allow evaluation of functions on uncertain values
