### 2019-09-26 efb3919
Always promote uncertain populations to float
