### 2019-09-26 231df8a
Constrain a strictly numeric uncertain population
