+++ src/sampling_constraints/truncation/truncate_population.jl
@@ -1,11 +1,12 @@
 import ..UncertainValues:
-    AbstractScalarPopulation
+    AbstractScalarPopulation,
+    UncertainScalarPopulation
 
 function Base.truncate(p::AbstractScalarPopulation, constraint::NoConstraint)
     p.values, p.probs
 end
 
-function Base.truncate(p::AbstractScalarPopulation, constraint::TruncateRange)
+function Base.truncate(p::UncertainScalarPopulation{T, PW}, constraint::TruncateRange) where {T <: Number, PW}
     inds = findall(constraint.min .<= p.values .<= constraint.max)
 
     if length(inds) == 0 
@@ -15,7 +16,7 @@ function Base.truncate(p::AbstractScalarPopulation, constraint::TruncateRange)
     p.values[inds], p.probs[inds]
 end
 
-function Base.truncate(p::AbstractScalarPopulation, constraint::TruncateMaximum)
+function Base.truncate(p::UncertainScalarPopulation{T, PW}, constraint::TruncateMaximum) where {T <: Number, PW}
     inds = findall(p.values .<= constraint.max)
 
     if length(inds) == 0 
@@ -26,7 +27,7 @@ function Base.truncate(p::AbstractScalarPopulation, constraint::TruncateMaximum)
 end
 
 
-function Base.truncate(p::AbstractScalarPopulation, constraint::TruncateMinimum)
+function Base.truncate(p::UncertainScalarPopulation{T, PW}, constraint::TruncateMinimum) where {T <: Number, PW}
     inds = findall(constraint.min .<= p.values)
 
     if length(inds) == 0 
@@ -37,7 +38,7 @@ function Base.truncate(p::AbstractScalarPopulation, constraint::TruncateMinimum)
 end
 
 
-function Base.truncate(p::AbstractScalarPopulation, constraint::TruncateLowerQuantile)
+function Base.truncate(p::UncertainScalarPopulation{T, PW}, constraint::TruncateLowerQuantile) where {T <: Number, PW}
     lower_bound = quantile(p, constraint.lower_quantile)
 
     inds = findall(lower_bound .<= p.values)
@@ -49,7 +50,7 @@ function Base.truncate(p::AbstractScalarPopulation, constraint::TruncateLowerQua
     p.values[inds], p.probs[inds]
 end
 
-function Base.truncate(p::AbstractScalarPopulation, constraint::TruncateUpperQuantile)
+function Base.truncate(p::UncertainScalarPopulation{T, PW}, constraint::TruncateUpperQuantile) where {T <: Number, PW}
     upper_bound = quantile(p, constraint.upper_quantile)
 
     inds = findall(p.values .<= upper_bound)
@@ -62,7 +63,7 @@ function Base.truncate(p::AbstractScalarPopulation, constraint::TruncateUpperQua
 end
 
 
-function Base.truncate(p::AbstractScalarPopulation, constraint::TruncateQuantiles)
+function Base.truncate(p::UncertainScalarPopulation{T, PW}, constraint::TruncateQuantiles) where {T <: Number, PW}
     lower_bound = quantile(p, constraint.lower_quantile)
     upper_bound = quantile(p, constraint.upper_quantile)
 
@@ -76,7 +77,7 @@ function Base.truncate(p::AbstractScalarPopulation, constraint::TruncateQuantile
 end
 
 
-function Base.truncate(p::AbstractScalarPopulation, constraint::TruncateStd; n::Int = 30000)
+function Base.truncate(p::UncertainScalarPopulation{T, PW}, constraint::TruncateStd; n::Int = 30000) where {T <: Number, PW}
     p_mean = mean(p, n)
     p_stdev = std(p, n)
     nσ  = constraint.nσ
[…]
