### 2019-09-26 7c49fa3
remove code that belongs in `Statistics` submodule
