+++ src/uncertain_values/abstract_types/AbstractScalarPopulation.jl
@@ -31,13 +31,4 @@ Base.minimum(p::AbstractScalarPopulation) = minimum(p.values)
 Base.maximum(p::AbstractScalarPopulation) = maximum(p.values)
 Distributions.support(p::AbstractScalarPopulation) = interval(minimum(p), maximum(p))
 
-StatsBase.mean(p::AbstractScalarPopulation, n::Int = 30000) = mean(rand(p, n))
-StatsBase.median(p::AbstractScalarPopulation, n::Int = 30000) = median(rand(p, n))
-StatsBase.middle(p::AbstractScalarPopulation, n::Int = 30000) = middle(rand(p, n))
-StatsBase.quantile(p::AbstractScalarPopulation, q, n::Int = 30000) = quantile(rand(p, n), q)
-StatsBase.std(p::AbstractScalarPopulation, n::Int = 30000) = std(rand(p, n))
-
-
-
-
 export AbstractScalarPopulation
