### 2019-09-26 68c91de
Update UncertainScalarPopulation constructors
