### 2019-09-26 03e9164
Test truncation of UncertainScalarPopulation
