### 2019-09-26 eaf9b72
Test truncation of UncertainScalarPopulation
