### 2019-09-26 fd392dd
Test point estimates of statistics
