### 2019-09-26 ee8efed
Test pairwise estimates of statistics
