### 2019-09-26 5c56216
Added some example uncertain values
