### 2019-09-27 a84429b
Docs improvements
