### 2019-09-27 89912b8
Docs improvements
