### 2019-09-27 b59ae1a
Fix some point statistics
