### 2019-09-27 0604c44
Update tests for pointwise and pairwise statistics
