+++ test/uncertain_statistics/test_UncertainStatistics.jl
@@ -1,3 +1,0 @@
-include("test_core_stats.jl")
-include("test_hypothesistests.jl")
-include("test_hypothesistests_timeseries.jl")
