### 2019-09-27 97dde17
The package is now integrated with CausalityTools
