### 2019-09-27 kahaaga/UncertainData.jl#93 @kahaaga
Added in #86, along with many other point estimators.
