### 2019-09-27 58c44c0
More statistics on uncertain values and datasets
