+++ src/statistics/UncertainStatistics.jl
@@ -15,7 +15,8 @@ using Reexport
     include("statsbase/uncertain_values/core_stats_point_estimates.jl")
     include("statsbase/uncertain_values/core_stats_pair_estimates.jl")
 	include("statsbase/uncertain_datasets/core_stats_uncertaindatasets.jl")
-
+    include("statsbase/uncertain_datasets/core_stats_uncertaindatasets_single_estimates.jl")
+    
     include("hypothesis_tests/mann_whitney.jl")
     include("hypothesis_tests/t_tests.jl")
     include("hypothesis_tests/anderson_darling.jl")
