### 2019-09-27 b5bfce1
More statistics on uncertain values and datasets
