+++ src/statistics/statsbase/uncertain_datasets/core_stats_uncertaindatasets_single_estimates.jl
@@ -1,0 +1,615 @@
+import StatsBase
+
+export UVAL_COLLECTION_TYPES
+
+const UVAL_COLLECTION_TYPES = Union{UD, UV} where {
+    UD <: AbstractUncertainValueDataset, 
+    UV <: AbstractVector{T} where {
+        T <: AbstractUncertainValue}}
+
+"""
+    median(x::Union{AbstractUncertainValueDataset, 
+                    Vector{AbstractUncertainValue}, n::Int)
+
+Obtain a distribution for the median of a collection of uncertain values.
+This is done by first drawing `n` length-`L` realisations of `x`, where 
+`L = length(x)`. Then, the median is computed for each of those length-`L`
+realisations, yielding a distribution of median estimates. 
+
+Detailed steps:
+
+1. First, draw a length-`L` realisation of `x` by drawing one random 
+    number from  each uncertain value furnishing the dataset. The draws are 
+    independent, so that no element-wise dependencies (e.g. sequential
+    correlations) that are not already present in the data are introduced in 
+    the realisation.
+2. Compute the median for the realisation.
+3. Repeat the procedure `n` times, drawing `n` independent realisations of `x`.
+    This yields `n` estimates of the median of `x`, which is returned as 
+    a vector.
+"""
+function StatsBase.median(x::UVAL_COLLECTION_TYPES, n::Int) 
+    resample(StatsBase.median, x, n)
+end
+
+"""
+    mean(x::Union{AbstractUncertainValueDataset, 
+                    Vector{AbstractUncertainValue}, n::Int)
+
+Obtain a distribution for the mean of a collection of uncertain values.
+This is done by first drawing `n` length-`L` realisations of `x`, where 
+`L = length(x)`. Then, the mean is computed for each of those length-`L`
+realisations, yielding a distribution of mean estimates. 
+
+Detailed steps:
+
+1. First, draw a length-`L` realisation of `x` by drawing one random 
+    number from  each uncertain value furnishing the dataset. The draws are 
+    independent, so that no element-wise dependencies (e.g. sequential
+    correlations) that are not already present in the data are introduced in 
+    the realisation.
+2. Compute the mean for the realisation.
+3. Repeat the procedure `n` times, drawing `n` independent realisations of `x`.
+    This yields `n` estimates of the mean of `x`, which is returned as 
+    a vector.
+"""
+function StatsBase.mean(x::UVAL_COLLECTION_TYPES, n::Int) 
+    resample(StatsBase.mean, x, n)
+end
+
+"""
+    mode(x::Union{AbstractUncertainValueDataset, 
+                    Vector{AbstractUncertainValue}, n::Int)
+
+Obtain a distribution for the mode of a collection of uncertain values.
+This is done by first drawing `n` length-`L` realisations of `x`, where 
+`L = length(x)`. Then, the mode is computed for each of those length-`L`
+realisations, yielding a distribution of mode estimates. 
+
+Detailed steps:
+
+1. First, draw a length-`L` realisation of `x` by drawing one random 
+    number from  each uncertain value furnishing the dataset. The draws are 
+    independent, so that no element-wise dependencies (e.g. sequential
+    correlations) that are not already present in the data are introduced in 
+    the realisation.
+2. Compute the mode for the realisation, which is a vector of length `L`
+3. Repeat the procedure `n` times, drawing `n` independent realisations of `x`.
+    This yields `n` estimates of the mode of `x`, which is returned as 
+    a vector.
+"""
+function StatsBase.mode(x::UVAL_COLLECTION_TYPES, n::Int) 
+    resample(StatsBase.mode, x, n)
+end
+
+"""
+    quantile(x::Union{AbstractUncertainValueDataset, 
+                    Vector{AbstractUncertainValue}, q, n::Int)
+
[…]
