### 2019-09-27 d5e330c
Documentation fixes
