+++ docs/src/mathematics/elementary_operations.md
@@ -3,10 +3,10 @@
 Elementary mathematical operations (`+`, `-`, `*`, and `/`) between arbitrary 
 uncertain values of different types and scalars are supported. 
 
-### Syntax
-Because elementary operations should work on arbitrary uncertain values, a resampling 
-approach is used to perform the mathematical operations. All mathematical 
-operations thus return a vector containing the results of repeated element-wise operations 
+## Syntax
+
+Resampling is used to perform the mathematical operations. All mathematical 
+operations return a vector containing the results of repeated element-wise operations 
 (where each element is a resampled draw from the furnishing distribution(s) of the 
 uncertain value(s)). 
 
@@ -14,81 +14,78 @@ The default number of realizations is set to `10000`. This allows calling `uval1
 for two uncertain values `uval1` and `uval2`. If you need to tune the number of resample 
 draws to `n`, use the `+(uval1, uval2, n)` syntax. 
 
-### Future improvements
+## Future improvements
+
 In the future, elementary operations might be improved for certain combinations of uncertain 
 values where exact expressions for error propagation are now, for example using the 
 machinery in `Measurements.jl` for normally distributed values.
 
+## Supported operations
 
-# Supported operations 
+## Addition
 
-## Addition 
-
-
-```@docs 
+```@docs
 Base.:+(a::AbstractUncertainValue, b::AbstractUncertainValue)
 Base.:+(a::AbstractUncertainValue, b::AbstractUncertainValue, n::Int)
 ```
 
-```@docs 
+```@docs
 Base.:+(a::Real, b::AbstractUncertainValue)
 Base.:+(a::Real, b::AbstractUncertainValue, n::Int)
 ```
 
-```@docs 
+```@docs
 Base.:+(a::AbstractUncertainValue, b::Real)
 Base.:+(a::AbstractUncertainValue, b::Real, n::Int)
 ```
 
-
 ## Subtraction
 
-```@docs 
+```@docs
 Base.:-(a::AbstractUncertainValue, b::AbstractUncertainValue)
 Base.:-(a::AbstractUncertainValue, b::AbstractUncertainValue, n::Int)
 ```
 
-```@docs 
+```@docs
 Base.:-(a::Real, b::AbstractUncertainValue)
 Base.:-(a::Real, b::AbstractUncertainValue, n::Int)
 ```
 
-```@docs 
+```@docs
 Base.:-(a::AbstractUncertainValue, b::Real)
 Base.:-(a::AbstractUncertainValue, b::Real, n::Int)
 ```
 
-
 ## Multiplication
 
-```@docs 
+```@docs
 Base.:*(a::AbstractUncertainValue, b::AbstractUncertainValue)
 Base.:*(a::AbstractUncertainValue, b::AbstractUncertainValue, n::Int)
 ```
 
-```@docs 
+```@docs
 Base.:*(a::Real, b::AbstractUncertainValue)
 Base.:*(a::Real, b::AbstractUncertainValue, n::Int)
 ```
 
-```@docs 
+```@docs
 Base.:*(a::AbstractUncertainValue, b::Real)
 Base.:*(a::AbstractUncertainValue, b::Real, n::Int)
 ```
 
 ## Division
 
-```@docs 
+```@docs
 Base.:/(a::AbstractUncertainValue, b::AbstractUncertainValue)
 Base.:/(a::AbstractUncertainValue, b::AbstractUncertainValue, n::Int)
 ```
 
-```@docs 
+```@docs
 Base.:/(a::Real, b::AbstractUncertainValue)
 Base.:/(a::Real, b::AbstractUncertainValue, n::Int)
 ```
 
-```@docs 
+```@docs
[…]
