### 2019-09-27 3490b5f
Tests for statistics
