### 2019-09-27 0cdfe58
Tests for statistics
