### 2019-09-27 1500cd4
More statistics on uncertain values and dataset, plus docs and tests
