+++ src/statistics/statsbase/uncertain_datasets/core_stats_uncertaindatasets_pairwise_estimates.jl
@@ -1,0 +1,503 @@
+"""
+    countne(x::Union{AbstractUncertainValueDataset, Vector{AbstractUncertainValue}},
+            y::Union{AbstractUncertainValueDataset, Vector{AbstractUncertainValue}}, 
+            n::Int)
+
+Estimate a `n`-member distribution on the number of indices at which the elements of 
+two collections of uncertain values are not equal. 
+
+This is done by repeating the following procedure `n` times:
+    
+1. Draw a length-`L` realisation of `x` by drawing one random 
+    number from  each uncertain value furnishing the dataset. The draws are 
+    independent, so that no element-wise dependencies (e.g. sequential
+    correlations) that are not already present in the data are introduced in 
+    the realisation.
+2. Draw a length-`L` realisation of `y` in the same manner.
+3. Count the number of indices at which the elements of the two length-`L`
+    draws are not equal.
+
+This yields `n` counts of non-equal values between `n` pairs of independent 
+realisations of `x` and `y`. The `n`-member distribution of nonequal-value counts 
+is returned as a vector.
+"""
+StatsBase.countne(x::UVAL_COLLECTION_TYPES, y::UVAL_COLLECTION_TYPES, n::Int) = 
+    resample(StatsBase.countne, x, y, n)
+
+"""
+    counteq(x::Union{AbstractUncertainValueDataset, Vector{AbstractUncertainValue}},
+            y::Union{AbstractUncertainValueDataset, Vector{AbstractUncertainValue}}, 
+            n::Int)
+
+Estimate a `n`-member distribution on the number of indices at which the elements of 
+two collections of uncertain values are equal.
+
+This is done by repeating the following procedure `n` times:
+    
+1. First, draw a length-`L` realisation of `x` by drawing one random 
+    number from  each uncertain value furnishing the dataset. The draws are 
+    independent, so that no element-wise dependencies (e.g. sequential
+    correlations) that are not already present in the data are introduced in 
+    the realisation.
+2. Second, draw a length-`L` realisation of `y` in the same manner.
+3. Count the number of indices at which the elements of the two length-`L`
+    draws are equal.
+
+This yields `n` counts of non-equal values between `n` pairs of independent 
+realisations of `x` and `y`. The `n`-member distribution of equal-value counts 
+is returned as a vector.
+"""
+StatsBase.counteq(x::UVAL_COLLECTION_TYPES, y::UVAL_COLLECTION_TYPES, n::Int) = 
+    resample(StatsBase.counteq, x, y, n)
+
+"""
+    corkendall(x::Union{AbstractUncertainValueDataset, Vector{AbstractUncertainValue}},
+            y::Union{AbstractUncertainValueDataset, Vector{AbstractUncertainValue}}, 
+            n::Int)
+
+Estimate a `n`-member distribution on Kendalls's rank correlation 
+coefficient between two collections of uncertain values.
+
+This is done by repeating the following procedure `n` times:
+    
+1. First, draw a length-`L` realisation of `x` by drawing one random 
+    number from  each uncertain value furnishing the dataset. The draws are 
+    independent, so that no element-wise dependencies (e.g. sequential
+    correlations) that are not already present in the data are introduced in 
+    the realisation.
+2. Second, draw a length-`L` realisation of `y` in the same manner.
+3. Compute Kendall's rank correlation coefficient between the two length-`L`
+    draws.
+
+This yields `n` computations of Kendall's rank correlation coefficient 
+between `n` independent pairs of realisations of `x` and `y`. The 
[…]
