### 2019-09-27 50895ea
More statistics on uncertain values and dataset, plus docs and tests
