### 2019-09-27 kahaaga/UncertainData.jl#94 @kahaaga
Fixed in #86
