### 2019-09-27 94854da
Some basic stats for CertainValue
