### 2019-09-27 df93cd3
`constrain` for UncertainScalarPopulation
