### 2019-09-27 0a6b65c
minimum and maximum for UncertainScalarPopulation
