+++ src/uncertain_values/UncertainScalarPopulation.jl
@@ -115,6 +115,19 @@ function ConstrainedUncertainScalarPopulation(values::VT, probabilities) where V
     ConstrainedUncertainScalarPopulation(UncertainValue.(values), StatsBase.weights(probabilities))
 end
 
+
+Base.minimum(pop::UncertainScalarPopulation{T, PW} where {T <: AbstractUncertainValue, PW}) = 
+    minimum([minimum(uv) for uv in pop])
+
+Base.maximum(pop::UncertainScalarPopulation{T, PW} where {T <: AbstractUncertainValue, PW}) = 
+    maximum([maximum(uv) for uv in pop])
+
+Base.minimum(pop::UncertainScalarPopulation{T, PW} where {T <: Number, PW}) = 
+    minimum([minimum(uv) for uv in pop])
+
+Base.maximum(pop::UncertainScalarPopulation{T, PW} where {T <: Number, PW}) = 
+    maximum([maximum(uv) for uv in pop])
+
 export 
 UncertainScalarPopulation,
 ConstrainedUncertainScalarPopulation
