### 2019-09-27 e229a4e
Additional `resample` methods
