### 2019-09-27 5f55e0f
Truncating populations
