+++ src/sampling_constraints/truncation/truncate_population.jl
@@ -2,179 +2,418 @@ import ..UncertainValues:
     AbstractScalarPopulation,
     UncertainScalarPopulation
 
-function Base.truncate(p::AbstractScalarPopulation, constraint::NoConstraint)
-    p.values, p.probs
+"""
+    truncate(population::AbstractScalarPopulation, constraint::NoConstraint)
+
+Get the elements and the associated sampling weights of the `population` members 
+satisfying the sampling `constraint`.
+
+- If `constraint` is a `NoConstraint` instance, then all members and weights are 
+    returned unmodified.
+"""
+function Base.truncate(pop::AbstractScalarPopulation, constraint::NoConstraint)
+    ConstrainedUncertainScalarPopulation(pop.values, pop.probs)
 end
 
-function Base.truncate(p::UncertainScalarPopulation{T, PW}, constraint::TruncateRange) where {T <: Number, PW}
-    inds = findall(constraint.min .<= p.values .<= constraint.max)
+############################################################
+# Populations whose members are strictly real-valued scalars
+############################################################
+function Base.truncate(pop::UncertainScalarPopulation{T, PW}, constraint::TruncateRange) where {T <: Number, PW}
+    inds = findall(constraint.min .<= pop.values .<= constraint.max)
 
     if length(inds) == 0 
-        throw(ArgumentError("$p could not be truncated. No values left after truncation."))
+        throw(ArgumentError("$pop could not be truncated. No values left after truncation."))
     end
 
-    p.values[inds], p.probs[inds]
+    ConstrainedUncertainScalarPopulation(pop.values[inds], pop.probs[inds])
 end
 
-function Base.truncate(p::UncertainScalarPopulation{T, PW}, constraint::TruncateMaximum) where {T <: Number, PW}
-    inds = findall(p.values .<= constraint.max)
+function Base.truncate(pop::UncertainScalarPopulation{T, PW}, constraint::TruncateMaximum) where {T <: Number, PW}
+    inds = findall(pop.values .<= constraint.max)
 
     if length(inds) == 0 
-        throw(ArgumentError("$p could not be truncated. No values left after truncation."))
+        throw(ArgumentError("$pop could not be truncated. No values left after truncation."))
     end
 
-    p.values[inds], p.probs[inds]
+    ConstrainedUncertainScalarPopulation(pop.values[inds], pop.probs[inds])
 end
 
 
-function Base.truncate(p::UncertainScalarPopulation{T, PW}, constraint::TruncateMinimum) where {T <: Number, PW}
-    inds = findall(constraint.min .<= p.values)
+function Base.truncate(pop::UncertainScalarPopulation{T, PW}, constraint::TruncateMinimum) where {T <: Number, PW}
+    inds = findall(constraint.min .<= pop.values)
 
     if length(inds) == 0 
-        throw(ArgumentError("$p could not be truncated. No values left after truncation."))
+        throw(ArgumentError("$pop could not be truncated. No values left after truncation."))
     end
 
-    p.values[inds], p.probs[inds]
+    ConstrainedUncertainScalarPopulation(pop.values[inds], pop.probs[inds])
 end
 
 
-function Base.truncate(p::UncertainScalarPopulation{T, PW}, constraint::TruncateLowerQuantile) where {T <: Number, PW}
-    lower_bound = quantile(p, constraint.lower_quantile)
+function Base.truncate(pop::UncertainScalarPopulation{T, PW}, constraint::TruncateLowerQuantile) where {T <: Number, PW}
+    lower_bound = quantile(pop, constraint.lower_quantile)
 
-    inds = findall(lower_bound .<= p.values)
[…]
