### 2019-09-27 ea9fbe7
Documentation improvement
