### 2019-09-29 8db4981
Improved subtyping for theoretical distributions
