### 2019-09-29 a5b62d3
Improved subtyping for theoretical distributions
