+++ src/uncertain_values/UncertainScalarsTheoretical.jl
@@ -23,9 +23,9 @@ Statistics.var(fd::TheoreticalDistributionScalarValue) = var(fd.distribution)
 StatsBase.mode(fd::TheoreticalDistributionScalarValue) = mode(fd.distribution)
 
 
-abstract type AbstractUncertainOneParameterScalarValue <: TheoreticalDistributionScalarValue end
-abstract type AbstractUncertainTwoParameterScalarValue <: TheoreticalDistributionScalarValue end
-abstract type AbstractUncertainThreeParameterScalarValue <: TheoreticalDistributionScalarValue end
+abstract type AbstractUncertainOneParameterScalarValue{S <: ValueSupport, T1} <: TheoreticalDistributionScalarValue end
+abstract type AbstractUncertainTwoParameterScalarValue{S <: ValueSupport, T1, T2} <: TheoreticalDistributionScalarValue end
+abstract type AbstractUncertainThreeParameterScalarValue{S <: ValueSupport, T1, T2, T3} <: TheoreticalDistributionScalarValue end
 
 """
     struct ConstrainedUncertainScalarValueOneParameter{S, T1 <: Number}
@@ -41,7 +41,7 @@ where the original distribution has been truncated.
 - **`a`**: The original value of the parameter of the original distribution.
 """
 struct ConstrainedUncertainScalarValueOneParameter{S,
-        T1 <: Number} <: AbstractUncertainOneParameterScalarValue
+        T1 <: Number} <: AbstractUncertainOneParameterScalarValue{S, T1}
     distribution::Distribution{Univariate, S}
     a::T1
 end
@@ -62,7 +62,7 @@ where the original distribution has been truncated.
 - **`b`**: The original value of the second parameter of the original distribution.
 """
 struct ConstrainedUncertainScalarValueTwoParameter{S, T1 <: Number,
-        T2 <: Number} <: AbstractUncertainTwoParameterScalarValue
+        T2 <: Number} <: AbstractUncertainTwoParameterScalarValue{S, T1, T2}
     distribution::Distribution{Univariate, S}
     a::T1
     b::T2
@@ -86,7 +86,7 @@ where the original distribution has been truncated.
 - **`c`**: The original value of the third parameter of the original distribution.
 
 """
-struct ConstrainedUncertainScalarValueThreeParameter{S, T1 <: Number, T2 <: Number, T3 <: Number} <: AbstractUncertainTwoParameterScalarValue
+struct ConstrainedUncertainScalarValueThreeParameter{S, T1 <: Number, T2 <: Number, T3 <: Number} <: AbstractUncertainThreeParameterScalarValue{S, T1, T2, T3}
     distribution::Distribution{Univariate, S}
     a::T1
     b::T2
@@ -106,8 +106,7 @@ import Distributions.Frechet
 """
 Uncertain value represented by a generic three-parameter distribution.
 """
-struct UncertainScalarTheoreticalThreeParameter{T1<:Number, T2<:Number, T3<:Number,
-        S<:ValueSupport} <: AbstractUncertainThreeParameterScalarValue
+struct UncertainScalarTheoreticalThreeParameter{S<:ValueSupport, T1<:Number, T2<:Number, T3<:Number} <: AbstractUncertainThreeParameterScalarValue{S, T1, T2, T3}
     distribution::Distribution{Univariate, S}
     a::T1
     b::T2
@@ -117,8 +116,7 @@ end
 """
 Uncertain value represented by a generic two-parameter distribution.
 """
[…]
