### 2019-09-29 387e216
`minimum` and `maximum` methods for population to abstract type
