### 2019-09-29 c3c372b
Define `rand` for populations
