### 2019-09-29 ef300d4
Add some useful methods to CertainValue
