### 2019-09-29 000c878
Improved documentation for `TruncateStd`
