+++ docs/src/index.md
@@ -19,27 +19,39 @@ should live in the probability domain, not as single value representations of th
 In this package, uncertain data values are thus 
 [stored as probability distributions](uncertain_values/uncertainvalues_overview.md). 
 Only when performing a computation or plotting, the uncertain values are realized by 
-resampling the probability distributions furnishing them. 
+resampling the probability distributions furnishing them.
 
 ## Organising uncertain data
 
-Individual uncertain observations may be collected in 
-[UncertainDatasets](uncertain_datasets/uncertain_datasets_overview.md), which can be 
-sampled according to user-provided sampling constraints. Likewise, indices (e.g. time, 
-depth or any other index) of observations can also be represented as probability 
-distributions and may also sampled using constraints. 
+Individual uncertain observations of different types are seamlessly mixed and can
+be organised in [collections of uncertain values](uncertain_datasets/uncertain_datasets_overview.md).
 
-The [UncertainIndexValueDataset](uncertain_datasets/uncertain_indexvalue_dataset.md) type 
-allows you to work with datasets where both the 
-[indices](uncertain_datasets/uncertain_index_dataset.md) and the 
-[data values](uncertain_datasets/uncertain_value_dataset.md) are uncertain.
-This may be useful when you, for example, want to draw realizations of your dataset while 
-simultaneously enforcing 
-[sequential resampling](resampling/sequential/resampling_uncertaindatasets_sequential.md), 
-for example 
-[strictly increasing](resampling/sequential/resampling_indexvalue_sequential.md) age models.
+- The [UncertainValueDataset](uncertain_datasets/uncertain_value_dataset.md) type is 
+    just a wrapper for a vector of uncertain values.
+- The [UncertainIndexDataset](uncertain_datasets/uncertain_index_dataset.md) type 
+    behaves just as [UncertainValueDataset](uncertain_datasets/uncertain_value_dataset.md), but has certain resampling methods such as [sequential resampling](resampling/sequential/resampling_uncertaindatasets_sequential) associated with them.
+- The [UncertainIndexValueDataset](uncertain_datasets/uncertain_indexvalue_dataset.md) 
+    type allows you to be explicit that you're working with datasets where both the 
+    [indices](uncertain_datasets/uncertain_index_dataset.md) and the 
+    [data values](uncertain_datasets/uncertain_value_dataset.md) are uncertain. 
+    This may be useful when you, for example, want to draw realizations of your 
+    dataset while simultaneously enforcing 
+    [sequential resampling](resampling/sequential/resampling_uncertaindatasets_sequential.md) 
+    models. One example is resampling while ensuring the draws have 
+    [strictly increasing](resampling/sequential/resampling_indexvalue_sequential.md) 
+    age models.
 
-## Mathematical operations 
+## Resampling
+
+Resampling can be done by user-provided sampling constraints, either on the individual uncertain values, or even on the entire collections.
+
+- Individual uncertain values are resampled by drawing random numbers from their furnishing     distributions/populations. The common `resample` method does this for all uncertain value
+types. Resampling can either be done from the entire populations, or after first applying
+[sampling constraints](sampling_constraints/available_constraints.md) on the underlying distributions/populations.
+
+- [Collections of uncertain values](uncertain_datasets/uncertain_datasets_overview.md) may also be resampled. Resampling collections can be done assuming no sequential dependence for your data, or by applying sequential sampling models. During this process [sampling constraints](sampling_constraints/available_constraints.md) can be applied element-wise or on entire collections.
+
+## Mathematical operations
 
[…]
