### 2019-09-29 c4a8f16
Documentation improvements
