+++ docs/src/resampling/resampling_overview.md
@@ -1,18 +1,14 @@
 
-
 Because uncertain values are represented by 
 [some kind of probability distribution](../uncertain_values/uncertainvalues_overview.md),
-we may trivially resample them by drawing values from their furnishing distributions.
-
+we may trivially resample them by drawing random numbers from their furnishing distributions/populations.
 
 If needed, you may choose to 
-[constrain](../sampling_constraints/constrain_uncertain_values.md) an uncertain value 
-before resampling, using one of the available 
+[constrain](../sampling_constraints/constrain_uncertain_values.md) an uncertain value before resampling, using one of the available 
 [sampling constraints](../sampling_constraints/available_constraints.md).
 
-
-The `resample` function is used to resample uncertain values. For detailed instructions
-on how to sample uncertain values and datasets of uncertain values, see the following pages:
+The `resample` function is used to resample uncertain values. For detailed instructions on how to sample uncertain values and datasets of uncertain 
+values, see the following pages:
 
 - [Resampling uncertain values](resampling_uncertain_values.md)
 - [Resampling uncertain value datasets](resampling_uncertain_datasets.md)
