+++ docs/src/sampling_constraints/constrain_uncertain_values.md
@@ -1,11 +1,21 @@
-Uncertain values may be constrained in various ways, as visualised by the following example.
+# Constraining uncertain values
+
+## Documentation
+
+```@docs
+constrain(uv::AbstractUncertainValue, constraint::SamplingConstraint)
+```
 
 ![](constraining_uncertain_values.svg)
+
+## Quick examples
+
+Uncertain values may be constrained in various ways, as visualised by the following example.
 
 Which was generated as follows. Note that the plot recipe normalises the distributions
 after constraining the uncertain values.
 
-```julia 
+```julia
 uval = UncertainValue(Normal, 2, 5)
 
 p1 = plot(uval, label = "full support", title = "Quantile truncation")
@@ -21,15 +31,7 @@ plot!(constrain(uval, TruncateMinimum(-2)), ls = :dash, label = "minimum value t
 plot(p1, p2, layout = (2, 1), link = :x, xlabel = "value", ylabel = "probability")
 ```
 
-
-## Documentation
-
-```@docs
-constrain(uv::AbstractUncertainValue, constraint::SamplingConstraint)
-```
-
-
-## Examples
+## Constraining theoretical distributions
 
 ``` julia tab="Theoretical distribution"
 using UncertainData, Distributions
@@ -46,6 +48,8 @@ uvc_min = constrain(uv, TruncateMinimum(0.5))
 uvc_max = constrain(uv, TruncateMaximum(1.5))
 uvc_range = constrain(uv, TruncateRange(0.5, 1.5))
 ```
+
+## Constraining theoretical distributions with fitted parameters
 
 ``` julia tab="Theoretical distribution with fitted parameters"
 using UncertainData, Distributions
@@ -64,6 +68,8 @@ uvc_max = constrain(uv, TruncateMaximum(1.5))
 uvc_range = constrain(uv, TruncateRange(0.5, 1.5))
 ```
 
+## Constraining Kernel density estimated distributions
+
 ``` julia tab="Kernel density estimated distribution"
 # Define an uncertain value furnished by a kernel density estimate to the
 # distribution of the empirical data
@@ -78,3 +84,202 @@ uvc_min = constrain(uv, TruncateMinimum(13))
 uvc_max = constrain(uv, TruncateMaximum(13))
 uvc_range = constrain(uv, TruncateRange(11, 12))
 ```
+
+## Constraining (nested) weighted populations of uncertain values
+
+Let's define a complicated uncertain value that is defined by a nested weighted population.
+
+```julia
+# Some subpopulations consisting of both scalar values and distributions
+subpop1_members = [UncertainValue(Normal, 0, 1), UncertainValue(Uniform, -2, 2), -5]
+subpop2_members = [UncertainValue(Normal, -2, 1), UncertainValue(Uniform, -6, -1),
+                    -3, UncertainValue(Gamma, 1, 0.4)]
+
+# Define the probabilities of sampling the different population members within the 
+# subpopulations. Weights are normalised, so we can input any numbers here indicating 
+# relative importance
+subpop1_probs = [1, 2, 1]
+subpop2_probs = [0.1, 0.2, 0.3, 0.1]
+
+pop1 = UncertainValue(subpop1_members, subpop1_probs)
+pop2 = UncertainValue(subpop2_members, subpop2_probs)
+
+# Define the probabilities of sampling the two subpopulations in the overall population.
+pop_probs = [0.3, 0.7]
+
+# Construct overall population
+pop_mixed = UncertainValue([pop1, pop2], pop_probs)
+```
+
+Now we can draw samples from this nested population. Sampling directly from the 
+entire distribution is done by calling `resample(pop_mixed, n_draws)`. However,
[…]
