### 2019-09-29 444443e
Documentation improvements
