### 2019-09-29 26577db
`resample` for uncertain populations now just call `rand`
