### 2019-09-29 9f76579
Dispatch on more abstract methods for `constrain`
Remove redundant methods
