### 2019-09-29 936ec28
Gather abstract types and behaviour for theoretical distributions
