### 2019-09-29 974cd79
`rand` for TheoreticalDistributionScalarValue
