### 2019-09-29 5affd35
Add constructors
