### 2019-09-29 79240fa
Add constructors
