### 2019-09-29 b5badd1
Test new constructors
