+++ test/uncertain_values/test_uncertain_values.jl
@@ -11,6 +11,11 @@ y = UncertainValue([1, UncertainValue(Normal, 0, 1), 3], rand(3))
 ##################################################
 # Uncertain theoretical distributions
 ################################################
+
+@test UncertainValue(Normal(1, 2)) isa UncertainScalarNormallyDistributed
+@test UncertainValue(Normal(1, 2)) isa UncertainScalarNormallyDistributed
+@test UncertainValue(Truncated(Normal(1, 2), 0.2, 0.8)) isa ConstrainedUncertainScalarValueTwoParameter
+
 # Uncertain normally distributed values
 @test UncertainValue(Normal, 1, 0.2, nσ = 2, trunc_lower = -5) isa UncertainScalarNormallyDistributed
 @test UncertainValue(Normal, -3, 0.2) isa UncertainScalarNormallyDistributed
