+++ test/runtests.jl
@@ -5,7 +5,6 @@ using StaticArrays
 using StatsBase
 using KernelDensity
 
-
 #############################################
 # UncertainValues module
 #############################################
@@ -23,14 +22,63 @@ include("uncertain_values/populations/test_ConstrainedUncertainScalarPopulation.
 #############################################
 include("uncertain_datasets/test_uncertain_datasets.jl")
 
-
-
 #############################################
-# Resampling
+# Resampling uncertain values
 #############################################
 include("resampling/test_resampling_uncertain_values.jl")
+
+#############################################
+# Resampling uncertain tuples
+#############################################
 include("resampling/test_resampling_uncertain_tuples.jl")
-include("resampling/test_resampling_datasets.jl")
+
+#############################################
+# Resampling uncertain datasets
+#############################################
+
+# Define an array of uncertain values `uvals` that we can construct datasets from.
+include("resampling/define_uncertain_values.jl")
+
+# Resampling vectors of uncertain values
+#---------------------------------------
+include("resampling/uncertain_vectors/test_resampling_vectors.jl")
+include("resampling/uncertain_vectors/test_resampling_vectors_apply_funcs.jl")
+include("resampling/uncertain_vectors/test_resampling_vectors_constraints.jl")
+
+
+# Resampling uncertain datasets
+#-------------------------------
+include("resampling/uncertain_datasets/test_resampling_datasets.jl")
+
+# Resampling uncertain value datasets
+#-------------------------------------
+include("resampling/uncertain_datasets/test_resampling_datasets_uncertainvaluedataset.jl")
+include("resampling/uncertain_datasets/test_resampling_datasets_uncertainvaluedataset_apply_funcs.jl")
+include("resampling/uncertain_datasets/test_resampling_datasets_uncertainvaluedataset_constraints.jl")
+
+# Resampling uncertain index datasets
+#-------------------------------------
+include("resampling/uncertain_datasets/test_resampling_datasets_uncertainindexdataset.jl")
+include("resampling/uncertain_datasets/test_resampling_datasets_uncertainindexdataset_apply_funcs.jl")
+include("resampling/uncertain_datasets/test_resampling_datasets_uncertainindexdataset_constraints.jl")
+
+# Resampling uncertain index-value datasets
+#-------------------------------------
+include("resampling/uncertain_datasets/test_resampling_uncertainindexvaluedataset.jl")
+
+# Resampling uncertain vectors
+#-----------------------------
+include("resampling/uncertain_vectors/test_resampling_vectors.jl")
+
+# Special resampling constraints
+#-----------------------------
+include("resampling/uncertain_datasets/sequential/test_resampling_sequential_increasing.jl")
+include("resampling/uncertain_datasets/sequential/test_resampling_sequential_decreasing.jl")
+
+#############################################
+# Resampling uncertain datasets element-wise
+#############################################
+include("resampling/uncertain_datasets/test_resampling_abstractuncertainvaluedataset_elwise.jl")
 
 
 ############################################
@@ -56,23 +104,18 @@ include("uncertain_statistics/test_hypothesistests.jl")
 #include("uncertain_statistics/test_hypothesistests_timeseries.jl")
 
 
-#############################################
+#########################################
 # Resampling with constraints and models
-#############################################
+#########################################
 
-# #############################
+# #####################
 # # Resampling schemes
-# #############################
+# #####################
 # include("resampling/resampling_schemes/test_ConstrainedResampling.jl")
 
-#############################
-# Distribution truncation
-#############################
-#include("sampling_constraints/test_truncate_UncertainScalarPopulation.jl")
-
-# #############################
+# #######################
 # # Sampling constraints
-# #############################
[…]
