+++ src/sampling_constraints/constrain_certain_value.jl
@@ -1,10 +1,12 @@
 import ..UncertainValues: CertainValue
 
-Base.truncate(v::CertainValue) = v
-Base.truncate(v::CertainValue, s::SamplingConstraint) = v
 
 constrain(v::CertainValue) = v
 constrain(v::CertainValue, s::SamplingConstraint) = v
+constrain(v::CertainValue, s::TruncateLowerQuantile) = v
+constrain(v::CertainValue, s::TruncateUpperQuantile) = v
+constrain(v::CertainValue, s::TruncateQuantiles) = v
+constrain(v::CertainValue, s::TruncateStd) = v
 
 
 export constrain
