+++ test/resampling/uncertain_datasets/test_resampling_datasets_uncertainindexdataset.jl
@@ -1,5 +1,5 @@
 
-UI = UncertainIndexDataset(uidxs)
+UI = UncertainIndexDataset(example_uidxs)
 
 n = 3
 @test resample(UI) isa Vector{<:Real}
