+++ src/sampling_constraints/SamplingConstraints.jl
@@ -9,6 +9,9 @@ using Reexport
     # Fallbacks when encountering incompatible sampling constraints
     include("fallback_constraints.jl")
 
+    # Truncate values 
+    include("truncation/truncate.jl")
+    
     # Constrain uncertain values
     include("constrain_uncertainvalue.jl")
     include("constrain_certain_value.jl")

[…]
