+++ test/sampling_constraints/test_constrain_population.jl
@@ -5,6 +5,236 @@ pop1 = UncertainValue(
         [0.5, 0.5, 0.5, 0.5]
     )
 
+pop2 = UncertainValue(1:10 |> collect, rand(10))
+pop3 = UncertainValue(1.0:10.0 |> collect, Weights(rand(10)))
+
+# Uncertain population consisting of uncertain populations and other stuff
+pop4 = UncertainValue([pop1, pop2], [0.1, 0.5])
+pop5 = UncertainValue([pop1, pop2, 2, UncertainValue(Normal, -2, 3)], Weights(rand(4)));
+
+uncertain_scalar_populations = [pop1, pop2, pop3, pop4, pop5];
+
+T = Union{Nothing, ConstrainedUncertainScalarPopulation}
+
+@test constrain(pop1, TruncateMinimum(0.2)) isa T
+@test constrain(pop1, TruncateMaximum(3.0)) isa T
+@test constrain(pop1, TruncateRange(0.0, 3.0)) isa T
+@test constrain(pop1, TruncateLowerQuantile(0.2)) isa T
+@test constrain(pop1, TruncateUpperQuantile(0.8)) isa T
+@test constrain(pop1, TruncateQuantiles(0.2, 0.8)) isa T
+@test constrain(pop1, TruncateStd(1)) isa T
+@test constrain(pop1, TruncateStd(0.5)) isa T
+
+# pop2.values = [1, 2, 3]
+@test constrain(pop2, TruncateMinimum(0.2)) isa T
+@test constrain(pop2, TruncateMaximum(2.5)) isa T
+@test constrain(pop2, TruncateRange(0.5, 2.5)) isa T
+@test constrain(pop2, TruncateLowerQuantile(0.2)) isa T
+@test constrain(pop2, TruncateUpperQuantile(0.8)) isa T
+@test constrain(pop2, TruncateQuantiles(0.2, 0.8)) isa T
+@test constrain(pop2, TruncateStd(1)) isa T
+@test constrain(pop2, TruncateStd(0.5)) isa T
+
+
+# pop3.values = [1.0, 2.0, 3.0]
+@test constrain(pop3, TruncateMinimum(0.2)) isa T
+@test constrain(pop3, TruncateMaximum(2.5)) isa T
+@test constrain(pop3, TruncateRange(0.5, 2.5)) isa T
+@test constrain(pop3, TruncateLowerQuantile(0.2)) isa T
+@test constrain(pop3, TruncateUpperQuantile(0.8)) isa T
+@test constrain(pop3, TruncateQuantiles(0.2, 0.8)) isa T
+@test constrain(pop3, TruncateStd(1)) isa T
+@test constrain(pop3, TruncateStd(0.5)) isa T
+
+@test constrain(pop4, TruncateMinimum(-10)) isa T
+@test constrain(pop4, TruncateMaximum(10)) isa T
+@test constrain(pop4, TruncateRange(-10, 10)) isa T
+@test constrain(pop4, TruncateLowerQuantile(0.2)) isa T
+@test constrain(pop4, TruncateUpperQuantile(0.8)) isa T
+@test constrain(pop4, TruncateQuantiles(0.2, 0.8)) isa T
+@test constrain(pop4, TruncateStd(1)) isa T
+@test constrain(pop4, TruncateStd(0.5)) isa T
+
+@test constrain(pop5, TruncateMinimum(-10)) isa T
+@test constrain(pop5, TruncateMaximum(10)) isa T
+@test constrain(pop5, TruncateRange(-10, 10)) isa T
+@test constrain(pop5, TruncateLowerQuantile(0.2)) isa T
+@test constrain(pop5, TruncateUpperQuantile(0.8)) isa T
+@test constrain(pop5, TruncateQuantiles(0.2, 0.8)) isa T
[…]
