+++ test/resampling/uncertain_vectors/test_resampling_vectors_constraints.jl
@@ -1,13 +1,13 @@
 
-constraints = [i % 2 == 0 ? TruncateStd(0.3) : TruncateQuantiles(0.1, 0.9) 
-    for i in 1:length(uvals)]
+constraints = [TruncateQuantiles(0.1, 0.9) 
+    for i in 1:length(example_uvals)] #i % 2 == 0 ? TruncateStd(0.3) : 
 
 # Draw one realisation
-@test resample(uvals) isa Vector{<:Real}
-@test resample(uvals, constraints[1]) isa Vector{<:Real}
-@test resample(uvals, constraints) isa Vector{<:Real}
+@test resample(example_uvals) isa Vector{<:Real}
+@test resample(example_uvals, constraints[1]) isa Vector{<:Real}
+@test resample(example_uvals, constraints) isa Vector{<:Real}
 
 # Draw multiple realisations
-@test resample(uvals, 5) isa Vector{Vector{T}} where T<:Real
-@test resample(uvals, constraints, 5) isa Vector{Vector{T}} where T<:Real
-@test resample(uvals, constraints[1], 5) isa Vector{Vector{T}} where T<:Real
+@test resample(example_uvals, 5) isa Vector{Vector{T}} where T<:Real
+@test resample(example_uvals, constraints, 5) isa Vector{Vector{T}} where T<:Real
+@test resample(example_uvals, constraints[1], 5) isa Vector{Vector{T}} where T<:Real
