+++ test/resampling/uncertain_datasets/test_resampling_datasets_uncertainvaluedataset_apply_funcs.jl
@@ -2,7 +2,7 @@
 ##########################################
 # Some uncertain datasets
 ##########################################
-UV = UncertainValueDataset(uvals)
+UV = UncertainValueDataset(example_uvals)
 
 ##########################################
 # Apply functions to datasets `n` times

[…]
