### 2019-10-08 6361b79
Lots of improvements to docs
