### 2019-10-08 faa079e
Lots of improvements to docs
