+++ docs/src/uncertain_values/merging.md
@@ -1,9 +1,89 @@
-
 Because all uncertainties are handled using a resampling approach, it is trivial to 
-[`combine`](@ref) uncertain values of different types into a single 
-uncertain value. 
+[`combine`](@ref) or merge uncertain values of different types into a single uncertain value.
 
-# Without weights 
+# Nomenclature
+
+Depending on your data, you may want to choose of one the following ways of 
+representing multiple uncertain values as one:
+
+- [Combining](@ref uncertainvalue_combine). An ensemble of uncertain 
+    values is represented as a weighted population. This approach is nice if you want 
+    to impose expert-opinion on the relative sampling probabilities of uncertain 
+    values in the ensemble, but still sample from the entire supports of each of the
+    furnishing values. This introduces no additional approximations besides what 
+    is already present at the moment you define your uncertain values.
+- [Merging](@ref uncertainvalue_merge). Multiple uncertain values are merged using 
+    a kernel density estimate to the overall distribution. This approach introduces 
+    approximations *beyond* what is present in the uncertain values when you define them.
+
+# [Combining uncertain values: the population approach](@id uncertainvalue_combine)
+
+**Combining** uncertain values is done by representing them as a weighted population
+of uncertain values, which is illustrated in the following example:
+
+```julia
+# Assume we have done some analysis and have three points whose uncertainties 
+# significantly overlap.
+v1 = UncertainValue(Normal(0.13, 0.52))
+v2 = UncertainValue(Normal(0.27, 0.42))
+v3 = UncertainValue(Normal(0.21, 0.61))
+
+# Give each value equal sampling probabilities and represent as a population
+pop = UncertainValue([v1, v2, v3], [1, 1, 1])
+
+# Let the values v1, v2 and v3 be sampled with probability ratios 1-2-3
+pop = UncertainValue([v1, v2, v3], [1, 2, 3])
+```
+
+![](figs/combining_uncertain_values.svg)
+
+This is not restricted to normal distributions! We can combine any type of 
+value in our population, even populations!
+
+```julia
+# Consider a population of normal distributions, and a gamma distribution
+v1 = UncertainValue(Normal(0.265, 0.52))
+v2 = UncertainValue(Normal(0.311, 0.15))
+v3 = UncertainValue([v1, v2], [2, 1])
+v4 = UncertainValue(Gamma(0.5, -1))
+pts = [v1, v4]
+wts = [2, 1]
+
+# New population is a nested population with unequal weights
+pop = UncertainValue(pts, wts)
+
+d1 = density(resample(pop, 20000), label = "population")
+
+d2 = plot()
+density!(d2, resample(pop[1], 20000), label = "v1")
+density!(d2, resample(pop[2], 20000), label = "v2")
+
+plot(d1, d2, layout = (2, 1), xlabel = "Value", ylabel = "Density", link = :x, xlims = (-2.5, 2.5))
+```
+
+![](figs/combining_uncertain_values_ex2.svg)
+
+This makes it possible treat an ensemble of uncertain values as a single uncertain value.
+
+With equal weights, this introduces no bias beyond what is present in the data, 
+because resampling is done from the full supports of each of the furnishing values. 
+Additional information on relative sampling probabilities, however, be it informed by
+expert opinion or quantative estimates, is easily incorporated by adjusting 
+the sampling weights.
+
[…]
