### 2019-10-08 cd2974d
Use UVAL_COLLECTION_TYPES
