+++ src/statistics/statsbase/uncertain_datasets/core_stats_uncertaindatasets_pairwise_estimates.jl
@@ -1,7 +1,5 @@
 """
-    countne(x::Union{AbstractUncertainValueDataset, Vector{AbstractUncertainValue}},
-            y::Union{AbstractUncertainValueDataset, Vector{AbstractUncertainValue}}, 
-            n::Int)
+    countne(x::UVAL_COLLECTION_TYPES, y::UVAL_COLLECTION_TYPES, n::Int)
 
 Estimate a `n`-member distribution on the number of indices at which the elements of 
 two collections of uncertain values are not equal. 
@@ -25,9 +23,7 @@ StatsBase.countne(x::UVAL_COLLECTION_TYPES, y::UVAL_COLLECTION_TYPES, n::Int) =
     resample(StatsBase.countne, x, y, n)
 
 """
-    counteq(x::Union{AbstractUncertainValueDataset, Vector{AbstractUncertainValue}},
-            y::Union{AbstractUncertainValueDataset, Vector{AbstractUncertainValue}}, 
-            n::Int)
+    counteq(x::UVAL_COLLECTION_TYPES, y::UVAL_COLLECTION_TYPES, n::Int)
 
 Estimate a `n`-member distribution on the number of indices at which the elements of 
 two collections of uncertain values are equal.
@@ -51,9 +47,7 @@ StatsBase.counteq(x::UVAL_COLLECTION_TYPES, y::UVAL_COLLECTION_TYPES, n::Int) =
     resample(StatsBase.counteq, x, y, n)
 
 """
-    corkendall(x::Union{AbstractUncertainValueDataset, Vector{AbstractUncertainValue}},
-            y::Union{AbstractUncertainValueDataset, Vector{AbstractUncertainValue}}, 
-            n::Int)
+    corkendall(x::UVAL_COLLECTION_TYPES, y::UVAL_COLLECTION_TYPES, n::Int)
 
 Estimate a `n`-member distribution on Kendalls's rank correlation 
 coefficient between two collections of uncertain values.
@@ -77,9 +71,7 @@ StatsBase.corkendall(x::UVAL_COLLECTION_TYPES, y::UVAL_COLLECTION_TYPES, n::Int)
     resample(StatsBase.corspearman, x, y, n)
 
 """
-    corspearman(x::Union{AbstractUncertainValueDataset, Vector{AbstractUncertainValue}},
-        y::Union{AbstractUncertainValueDataset, Vector{AbstractUncertainValue}}, 
-        n::Int)
+    corspearman(x::UVAL_COLLECTION_TYPES, y::UVAL_COLLECTION_TYPES, n::Int)
 
 Estimate a `n`-member distribution on Spearman's rank correlation 
 coefficient between two collections of uncertain values.
@@ -103,9 +95,7 @@ StatsBase.corspearman(x::UVAL_COLLECTION_TYPES, y::UVAL_COLLECTION_TYPES, n::Int
     resample(StatsBase.corspearman, x, y, n)
 
 """
-    cor(x::Union{AbstractUncertainValueDataset, Vector{AbstractUncertainValue}},
-        y::Union{AbstractUncertainValueDataset, Vector{AbstractUncertainValue}}, 
-        n::Int)
+    cor(x::UVAL_COLLECTION_TYPES, y::UVAL_COLLECTION_TYPES, n::Int)
 
 Estimate a distribution on Pearson's rank correlation coefficient between 
 two collections of uncertain values.
@@ -129,9 +119,7 @@ StatsBase.cor(x::UVAL_COLLECTION_TYPES, y::UVAL_COLLECTION_TYPES, n::Int) =
     resample(StatsBase.cor, x, y, n)
 
 """
[…]
