### 2019-10-08 a08caf3
Use UVAL_COLLECTION_TYPES
