+++ src/statistics/statsbase/uncertain_datasets/core_stats_uncertaindatasets_single_estimates.jl
@@ -1,15 +1,7 @@
 import StatsBase
 
-export UVAL_COLLECTION_TYPES
-
-const UVAL_COLLECTION_TYPES = Union{UD, UV} where {
-    UD <: AbstractUncertainValueDataset, 
-    UV <: AbstractVector{T} where {
-        T <: AbstractUncertainValue}}
-
 """
-    median(x::Union{AbstractUncertainValueDataset, 
-                    Vector{AbstractUncertainValue}, n::Int)
+    median(x::UVAL_COLLECTION_TYPES, n::Int)
 
 Obtain a distribution for the median of a collection of uncertain values.
 This is done by first drawing `n` length-`L` realisations of `x`, where 
@@ -33,8 +25,7 @@ function StatsBase.median(x::UVAL_COLLECTION_TYPES, n::Int)
 end
 
 """
-    mean(x::Union{AbstractUncertainValueDataset, 
-                    Vector{AbstractUncertainValue}, n::Int)
+    mean(x::UVAL_COLLECTION_TYPES, n::Int)
 
 Obtain a distribution for the mean of a collection of uncertain values.
 This is done by first drawing `n` length-`L` realisations of `x`, where 
@@ -58,8 +49,7 @@ function StatsBase.mean(x::UVAL_COLLECTION_TYPES, n::Int)
 end
 
 """
-    mode(x::Union{AbstractUncertainValueDataset, 
-                    Vector{AbstractUncertainValue}, n::Int)
+    mode(x::UVAL_COLLECTION_TYPES, n::Int)
 
 Obtain a distribution for the mode of a collection of uncertain values.
 This is done by first drawing `n` length-`L` realisations of `x`, where 
@@ -83,8 +73,7 @@ function StatsBase.mode(x::UVAL_COLLECTION_TYPES, n::Int)
 end
 
 """
-    quantile(x::Union{AbstractUncertainValueDataset, 
-                    Vector{AbstractUncertainValue}, q, n::Int)
+    quantile(x::UVAL_COLLECTION_TYPES, q, n::Int)
 
 Obtain a distribution for the quantile(s) `q` of a collection of uncertain values.
 This is done by first drawing `n` length-`L` realisations of `x`, where 
@@ -109,8 +98,7 @@ end
 
 
 """
-    iqr(x::Union{AbstractUncertainValueDataset, 
-                    Vector{AbstractUncertainValue}, n::Int)
+    iqr(x::UVAL_COLLECTION_TYPES, n::Int)
 
 Obtain a distribution for the interquartile range (IQR), i.e. the 75th 
 percentile minus the 25th percentile, of a collection of uncertain values.
@@ -135,8 +123,7 @@ function StatsBase.iqr(x::UVAL_COLLECTION_TYPES, n::Int)
 end
 
 """
-    middle(x::Union{AbstractUncertainValueDataset, 
-                    Vector{AbstractUncertainValue}, n::Int)
+    middle(x::UVAL_COLLECTION_TYPES, n::Int)
 
 Obtain a distribution for the middle of a collection of uncertain values.
 This is done by first drawing `n` length-`L` realisations of `x`, where 
@@ -160,8 +147,7 @@ function StatsBase.middle(x::UVAL_COLLECTION_TYPES, n::Int)
 end
 
 """
-    std(x::Union{AbstractUncertainValueDataset, 
-                    Vector{AbstractUncertainValue}, n::Int)
+    std(x::UVAL_COLLECTION_TYPES, n::Int)
 
 Obtain a distribution for the standard deviation of a collection of uncertain values.
 This is done by first drawing `n` length-`L` realisations of `x`, where 
[…]
