+++ src/resampling/uncertain_dataset/resample_uncertaindataset_value.jl
@@ -1,7 +1,8 @@
 
 import ..UncertainDatasets: 
 	AbstractUncertainValueDataset,
-	UncertainValueDataset
+	UncertainValueDataset,
+	UVAL_COLLECTION_TYPES
 
 import ..SamplingConstraints:
 	NoConstraint,
@@ -9,162 +10,167 @@ import ..SamplingConstraints:
 	TruncateUpperQuantile,
 	TruncateQuantiles
 
-function resample(uv::DT, constraint::Vector{<:SamplingConstraint}) where {
-		DT <: AbstractUncertainValueDataset}
-    [resample(uv.values[i], constraint[i]) for i in 1:length(uv)]
-end
+"""
+	resample(x::UVAL_COLLECTION_TYPES, constraint::SamplingConstraint) -> Vector{T} where T
 
-function resample(uv::DT, constraint::Vector{<:SamplingConstraint}, n::Int) where {
-	DT <: AbstractUncertainValueDataset}
-[[resample(uv.values[i], constraint[i]) for i in 1:length(uv)] for k = 1:n]
-end
+Resample `x` (a collection of uncertain values) once, applying the provided sampling `constraint`.
 
-function resample(uv::DT, constraint::SamplingConstraint) where {
-		DT <: AbstractUncertainValueDataset}
+Returns a `length(x)`-element vector. The `i`-th element of this vector is generated by 
+truncating the `i`-th uncertain value by the sampling `constraint`, then drawing a single random 
+number from the truncated value.
+
+See also [`UVAL_COLLECTION_TYPES`](@ref).
+
+## Example 
+
+```julia
+# Generate some uncertain values where the `i`-th value is given by a normal 
+# distribution with mean `i` and a standard deviation drawn from a uniform 
+# distribution on `[0, 1]`.
+uvals = [UncertainValue(Normal(i, rand())) for i = 1:100]
+
+# Truncate each distribution at +- 0.5 standard deviations, then resample. 
+resample(uvals, TruncateStd(0.5))
+```
+"""
+function resample(uv::UVAL_COLLECTION_TYPES, constraint::SamplingConstraint)
 	[resample(uv.values[i], constraint) for i in 1:length(uv)]
 end
 
-function resample(uv::DT, constraint::SamplingConstraint, n::Int) where {
-		DT <: AbstractUncertainValueDataset}
+"""
+	resample(x::UVAL_COLLECTION_TYPES, constraint::Vector{<:SamplingConstraint}) -> Vector{T} where T
+
+Resample `x` (a collection of uncertain values) once, applying the provided sampling `constraint`s.
+The number of constraints must match the number of elements in `x`.
+
+Returns a `length(x)`-element vector. The `i`-th element of this vector is generated by 
+truncating the `i`-th uncertain value by the `i`-th sampling `constraint`, then drawing a single random 
+number from the truncated value. 
+
+See also [`UVAL_COLLECTION_TYPES`](@ref).
+
+## Example
+
+```julia
+# Generate some uncertain values where the `i`-th value is given by a normal 
+# distribution with mean `i` and a standard deviation drawn from a uniform 
+# distribution on `[0, 1]`.
+uvals = [UncertainValue(Normal(i, rand())) for i = 1:100]
+
+# Truncate each distribution at +- 0.5 standard deviations, then resample. 
+resample(uvals, TruncateStd(0.5))
+```
+"""
+function resample(uv::UVAL_COLLECTION_TYPES, constraint::Vector{<:SamplingConstraint})
+    [resample(uv.values[i], constraint[i]) for i in 1:length(uv)]
+end
+
+"""
+	resample(x::UVAL_COLLECTION_TYPES, constraint::SamplingConstraint, n::Int) -> Vector{Vector{T}} where T
+
[…]
