### 2019-10-08 a9b33dd
Update docs for `CertainValue` constructor
