### 2019-10-08 b3b8104
Some documentation changes
