+++ src/uncertain_values/UncertainScalarsKDE.jl
@@ -19,7 +19,6 @@ An empirical value represented by a distribution estimated from actual data.
 - **`values`**: The values from which `distribution` is estimated.
 - **`range`**: The values for which the pdf is estimated.
 - **`pdf`**: The values of the pdf at each point in `range`.
-
 """
 struct UncertainScalarKDE{T} <: AbstractUncertainScalarKDE{T}
     distribution::KernelDensity.UnivariateKDE
@@ -32,7 +31,7 @@ end
 """
     TruncatedUncertainScalarKDE
 
-A truncated UncertainScalarKDE.
+A truncated [`UncertainScalarKDE`](@ref).
 """
 struct TruncatedUncertainScalarKDE{T} <: AbstractUncertainScalarKDE{T}
     distribution::KernelDensity.UnivariateKDE
