### 2019-10-08 b01f396
Some documentation changes
