+++ docs/src/publications.md
@@ -1,13 +1,14 @@
 
-## Scientific papers 
+## Scientific papers
+
 - [Vasskog, Kristian, John‐Inge Svendsen, Jan Mangerud, Kristian Agasøster Haaga, 
     Arve Svean, and Eva Maria Lunnan. "Evidence of early deglaciation (18 000 cal a bp) 
     and a postglacial relative sea‐level curve from southern Karmøy, south‐west Norway." 
     Journal of Quaternary Science 
     (2019)](https://onlinelibrary.wiley.com/doi/full/10.1002/jqs.3109).
 
+## Software
 
-## Software 
-- In a coming release, [CausalityTools.jl](https://github.com/kahaaga/CausalityTools.jl) 
+- [CausalityTools.jl](https://github.com/kahaaga/CausalityTools.jl) version >= 0.3.0
     uses UncertainData.jl to detect causal relationships between time series with 
-    uncertainties both in (time) indices and data values.
+    uncertainties.
