### 2019-10-08 e69c116
Add Measurements.jl to docs Project.toml
