+++ docs/Project.toml
@@ -4,3 +4,4 @@ DocumenterMarkdown = "997ab1e6-3595-5248-9280-8efb232c3433"
 Distributions = "31c24e10-a181-5473-b8eb-7969acd0382f"
 KernelDensity = "5ab0869b-81aa-558d-bb23-cbf5423bbe9b"
 StatsBase = "2913bbd2-ae8a-5f71-8c99-4fb6c76f3a91"
+Measurements = "eff96d63-e80a-5855-80a2-b1b0885c5ab7"
