### 2019-10-08 3f15310
Doc fixes
