### 2019-10-08 d4b3ae4
Doc fixes
