### 2019-10-08 7afdd84
Doc fixes
