### 2019-10-08 df92378
Docs fixes
