+++ docs/src/index.md
@@ -44,10 +44,11 @@ Statistics on uncertain datasets are computed using a resampling approach:
 
 [Resampling](resampling/resampling_overview.md) is done by drawing random numbers from the furnishing distributions/populations of the uncertain value(s), using one of the [`resample`](@ref) methods.
 
-- [Individual uncertain values](resampling/resampling_uncertain_values) may be sampled as they are,
-or after first applying [sampling constraints](sampling_constraints/available_constraints.md) on the underlying distributions/populations.
+- [Individual uncertain values](resampling/resampling_uncertain_values) may be sampled as they 
+    are, or after first applying [sampling constraints](sampling_constraints/available_constraints.md) on the underlying distributions/populations.
 
-- [Collections of uncertain values](resampling/resampling_uncertain_datasets.md) can be resampled by either assuming no sequential dependence for your data, or by applying sequential sampling models. During this process [sampling constraints](sampling_constraints/available_constraints.md) can be applied element-wise or on entire collections.
+- [Collections of uncertain values](resampling/resampling_uncertain_datasets.md) can be 
+    resampled by either assuming no sequential dependence for your data, or by applying sequential sampling models. During this process [sampling constraints](sampling_constraints/available_constraints.md) can be applied element-wise or on entire collections.
 
 ## Basic workflow
 
