+++ docs/src/uncertain_values/uncertainvalues_fitted.md
@@ -1,3 +1,5 @@
+# [Fitted theoretical distributions](@id uncertain_value_fitted_theoretical_distribution)
+
 For data values with histograms close to some known distribution, the user
 may choose to represent the data by fitting a theoretical distribution to the
 values. This will only work well if the histogram closely resembles a
