+++ docs/src/uncertain_values/uncertainvalues_populations.md
@@ -1,4 +1,6 @@
 
+# [Weighted populations](@id uncertain_value_population)
+
 The `UncertainScalarPopulation` type allows representation of an uncertain scalar 
 represented by a population of values who will be sampled according to a vector of 
 explicitly provided probabilities. Think of it as an explicit kernel density estimate. 
