### 2019-10-09 kahaaga/UncertainData.jl#92 @kahaaga
Closing this in favour of just having better documentation for the statistical functions, and adding a section of error propagation in the documentation.
