### 2019-10-09 #1666 @ahwillia
Looks good to me!

On Wed, Oct 9, 2019, 1:33 AM Kristian Agasøster Haaga <
notifications@github.com> wrote:

> I've now re-done the entire documentation for the package to make the
> points that @ahwillia <https://github.com/ahwillia> pointed out much more
> clear. Instead of making tutorials explaining the difference (which might
> be missed by users unless they really dive into the documentation), I now
> explain the resampling procedure in detail for each function that uses it.
>
> Concretely, this involved re-working the documentation for the statistical
> functions from scratch, adding detailed sections on how each of the
> computations are done. Moreover, I added a section on the difference
> between exact error propagation and approximate error propagation used in
> this package. I'll add tutorials later, but that is a bit beyond the scope
> of the software itself. It should be clear for the user what assumptions
> goes into the computations now.
>
> I hope that satisfactorily addresses your comments, @ahwillia
> <https://github.com/ahwillia>.
>
> As for the code, I've submitted version 0.5.0 (which incorporates the
> above changes and some more) to the Julia registry. If everything goes
> fine, the new version should be registered and tagged within a day or two.
> Once the process is finished, I'll tag you here, @oliviaguest
> <https://github.com/oliviaguest>.
>
