### 2019-10-10 446c068
Support CertainValues when resampling with strictly increasing/decreasing resampling
