### 2019-10-10 726f583
Support CertainValues when resampling with strictly increasing/decreasing resampling
