### 2019-10-10 #1666 @kahaaga
@oliviaguest The newest version of my package has been registered in the Julia repository and tagged with version 0.5.1 (see [release list](https://github.com/kahaaga/UncertainData.jl/releases)). I don't have more to add at this stage, so all should be ready for re-generating the pdf!
