### 2019-10-12 3ad763e
`UncertainValue` constructor for vectors of zero-dimensional arrays.
