### 2019-10-13 d7a00b4
Many new resampling schemes and resample methods that dispatch on them
