### 2019-10-13 949a9cb
Many new resampling schemes and resample methods that dispatch on them
