### 2019-10-13 6a1953a
Added missing `constrain` method
