### 2019-10-13 21ad107
Improve type documentation and include example
