### 2019-10-13 1c93aaa
Added tests, resample methods and docs
