+++ docs/src/resampling/resampling_schemes/resampling_with_schemes_uncertain_value_collections.md
@@ -1,0 +1,10 @@
+
+# [Resampling with schemes](@id applying_resampling_scheme_uncertain_value_collections)
+
+For some uncertain collections and datasets, special resampling types are available to make resampling easier.
+
+## Constrained resampling
+
+```@docs
+resample(::AbstractUncertainValueDataset, ::ConstrainedValueResampling{1})
+```
