+++ docs/src/resampling/resampling_overview.md
@@ -9,6 +9,15 @@ If needed, you may choose to
 The `resample` function is used to resample uncertain values. For detailed instructions on how to sample uncertain values and datasets of uncertain 
 values, see the following pages:
 
+# Resampling uncertain values
+
 - [Resampling uncertain values](resampling_uncertain_values.md)
-- [Resampling uncertain value datasets](resampling_uncertain_datasets.md)
-- [Resampling uncertain index-value datasets](resampling_uncertain_indexvalue_datasets.md)
+- [Resampling uncertain value datasets](resampling_uncertain_datasets.md). See also the
+    [resampling schemes](@ref resampling_schemes_uncertainvaluecollections) which can be
+    [applied](@ref applying_resampling_scheme_uncertain_value_collections) to 
+    simplify resampling.
+- [Resampling uncertain index-value datasets](resampling_uncertain_indexvalue_datasets.md).
+    See also the
+    [resampling schemes](@ref resampling_schemes_uncertainindexvaluecollections) which can be
+    [applied](@ref applying_resampling_scheme_uncertain_indexvalue_collections) to 
+    simplify resampling.
