### 2019-10-13 d1b89fd
Added tests, resample methods and docs
