+++ test/resampling/resampling_schemes/test_SequentialResampling.jl
@@ -1,0 +1,3 @@
+seq = SequentialResampling(StrictlyIncreasing())
+
+@test seq isa SequentialResampling
