+++ test/resampling/uncertain_datasets/test_resampling_with_schemes.jl
@@ -1,7 +1,6 @@
-###########################################################
-# SequentialResampling and SequentialInterpolatedResampling
-###########################################################
-
+################
+# Example data 
+################
 N = 50
 x_uncertain = [UncertainValue(Normal, x, rand(Uniform(0.1, 0.8))) for x in rand(N)]
 y_uncertain = [UncertainValue(Normal, y, rand(Uniform(0.1, 0.8))) for y in rand(N)]
@@ -16,71 +15,68 @@ timeinds_y = UncertainIndexDataset(time_certain)
 X = UncertainIndexValueDataset(timeinds_x, x)
 Y = UncertainIndexValueDataset(timeinds_y, y);
 
+###########################################################
+# SequentialResampling
+###########################################################
 seq = SequentialResampling(StrictlyIncreasing())
-seq_intp = SequentialInterpolatedResampling(StrictlyIncreasing(), RegularGrid(0, N, 2))
 
 @test resample(X, seq) isa NTuple{2, Vector{<:Real}}
 @test resample(X, seq)[1] |> length == N
 @test resample(X, seq)[2] |> length == N
-@test resample(X, seq_intp) isa Tuple{AbstractRange, Vector{<:Real}}
-@test resample(X, seq_intp)[1] |> length == length(0:2:N)
-@test resample(X, seq_intp)[2] |> length == length(0:2:N)
-
 @test resample(Y, seq) isa NTuple{2, Vector{<:Real}}
 @test resample(Y, seq)[1] |> length == N
 @test resample(Y, seq)[2] |> length == N
+
+###########################################################
+# SequentialInterpolatedResampling
+###########################################################
+seq_intp = SequentialInterpolatedResampling(StrictlyIncreasing(), RegularGrid(0, N, 2))
+
+@test resample(X, seq_intp) isa Tuple{AbstractRange, Vector{<:Real}}
+@test resample(X, seq_intp)[1] |> length == length(0:2:N)
+@test resample(X, seq_intp)[2] |> length == length(0:2:N)
 @test resample(Y, seq_intp) isa Tuple{AbstractRange, Vector{<:Real}}
 @test resample(Y, seq_intp)[1] |> length == length(0:2:N)
 @test resample(Y, seq_intp)[2] |> length == length(0:2:N)
 
-# Add uncertainties to the time series values
-x_uncertain = [UncertainValue(Normal, x, rand(Uniform(0.1, 0.8))) for x in rand(N)]
-y_uncertain = [UncertainValue(Normal, y, rand(Uniform(0.1, 0.8))) for y in rand(N)]
-x = UncertainValueDataset(x_uncertain)
-y = UncertainValueDataset(y_uncertain)
+#################################
+# ConstrainedValueResampling
+#################################
 
-time_uncertain = [UncertainValue(Normal, i, 1) for i = 1:length(x)];
-time_certain = [CertainValue(i) for i = 1:length(x)];
-timeinds_x = UncertainIndexDataset(time_uncertain)
-timeinds_y = UncertainIndexDataset(time_certain)
+c = ConstrainedValueResampling(10, TruncateStd(1))
 
-X = UncertainIndexValueDataset(timeinds_x, x)
-Y = UncertainIndexValueDataset(timeinds_y, y);
+r = resample(timeinds_y, c)
+@test length(r) == c.n
+@test length(r[1]) == N
+
+r = resample(timeinds_x, c)
+@test length(r) == c.n
+@test length(r[1]) == N
+
+r = resample(x, c)
+@test length(r) == c.n
+@test length(r[1]) == N
+
+r = resample(y, c)
[…]
