+++ docs/src/resampling/resampling_schemes/resampling_schemes_uncertain_value_collections.md
@@ -1,0 +1,9 @@
+# [List of resampling schemes and their purpose](@id resampling_schemes_uncertainvaluecollections)
+
+For collections of uncertain data, sampling constraints can be represented using the [`ConstrainedValueResampling`](@ref) type. This allows for passing complicated sampling constraints as a single input argument to functions that accept uncertain value collections.
+
+## Constrained resampling
+
+```@docs
+ConstrainedValueResampling
+```

[…]
