### 2019-10-18 91415d1
Resampling in-place
