+++ test/runtests.jl
@@ -83,6 +83,10 @@ include("resampling/uncertain_vectors/test_resampling_vectors.jl")
 include("resampling/uncertain_datasets/sequential/test_resampling_sequential_increasing.jl")
 include("resampling/uncertain_datasets/sequential/test_resampling_sequential_decreasing.jl")
 
+# Resampling inplace.
+#-----------------------------
+include("resampling/test_resampling_inplace.jl")
+
 #############################################
 # Resampling uncertain datasets element-wise
 #############################################
