+++ docs/src/resampling/resampling_inplace.md
@@ -1,0 +1,4 @@
+
+```@docs 
+resample!
+```
