+++ src/UncertainData.jl
@@ -81,6 +81,7 @@ include("uncertain_values/operations/merging.jl")
 
 # Example datasets 
 include("example_datasets/example_uvals.jl")
+include("example_datasets/example_uncertainindexvalue_dataset.jl")
 
 export UncertainScalarBinomialDistributed
 
