### 2019-10-18 e213bb0
Require DynamicalSystemsBase for trajectories of dynamical systems
