### 2019-10-18 04e9985
Test inplace resampling
