### 2019-10-18 f21a0df
Resample index-value datasets in-place.
