+++ test/resampling/test_resampling_inplace.jl
@@ -33,3 +33,35 @@ resample!(v, example_uvals)
 arr = fill(NaN, 100, length(example_uvals))
 resample!(arr, example_uvals)
 @test any(isnan.(arr)) == false
+
+
+#################################################
+# Uncertain index-value collections into vectors
+#################################################
+ids = UncertainIndexDataset(example_uidxs)
+vals = UncertainValueDataset(example_uvals)
+U = UncertainIndexValueDataset(ids, vals)
+
+idxs = fill(NaN, length(U))
+vals = fill(NaN, length(U))
+
+resample!(idxs, vals, U)
+
+@test any(isnan.(idxs)) == false
+@test any(isnan.(vals)) == false
+
+#################################################
+# Uncertain index-value collections into arrays
+#################################################
+ids = UncertainIndexDataset(example_uidxs)
+vals = UncertainValueDataset(example_uvals)
+U = UncertainIndexValueDataset(ids, vals)
+
+n_draws = 10
+idxs = fill(NaN, n_draws, length(U))
+vals = fill(NaN, n_draws, length(U))
+
+resample!(idxs, vals, U)
+
+@test any(isnan.(idxs)) == false
+@test any(isnan.(vals)) == false
