### 2019-10-19 e0a80dd
Release 0.7.0 (#111)
- Resampling in-place
- Added some example constraints.
