### 2019-10-22 af5d231
Improve docs and add tutorial on resampling with binning
