+++ docs/src/resampling/resampling_schemes/resampling_with_schemes_uncertain_indexvalue_collections.md
@@ -1,7 +1,9 @@
 
-# [Resampling with schemes](@id applying_resampling_scheme_uncertain_indexvalue_collections)
+# [Resampling schemes](@id applying_resampling_scheme_uncertain_indexvalue_collections)
 
 For some uncertain collections and datasets, special resampling types are available to make resampling easier.
+
+# Constrained resampling schemes
 
 ## Constrained resampling
 
@@ -9,14 +11,30 @@ For some uncertain collections and datasets, special resampling types are availa
 resample(::AbstractUncertainIndexValueDataset, ::ConstrainedIndexValueResampling{2, 1})
 ```
 
-## Sequential resampling
+# Sequential resampling schemes
+
+## Sequential
 
 ```@docs
 resample(::AbstractUncertainIndexValueDataset, ::SequentialResampling)
 ```
 
-## Sequential and interpolated resampling
+## Sequential and interpolated
 
 ```@docs
 resample(::AbstractUncertainIndexValueDataset, ::SequentialInterpolatedResampling)
+```
+
+# Binned resampling schemes
+
+## BinnedResampling
+
+```@docs
+resample(::AbstractUncertainIndexValueDataset, ::BinnedResampling)
+```
+
+## BinnedMeanResampling
+
+```@docs
+resample(x::AbstractUncertainIndexValueDataset, resampling::BinnedMeanResampling)
 ```
