+++ docs/src/tutorials/tutorial_transforming_data_to_regular_grid.md
@@ -1,0 +1,161 @@
+# [Transforming uncertain data to a regular grid](@id transform_data_to_regular_grid)
+
+Time series analysis algorithms often require data that are equally spaced in time. 
+Dealing with data that have uncertainties both in values and in time, that becomes tricky.
+A solution is to partition the time axis into bins of a certain size, transform 
+your data onto those bins, the compute your statistic on the transformed data.
+
+This tutorial shows how uncertain data can be transformed to a regular grid using a 
+combination of resampling and binning.
+
+## Some example data
+
+We'll look at the first and second variables of an autoregressive system with 
+unidirectional coupling. We'll use 100 points where each time point is spaced 
+10 time unit apart. In addition we'll make the positions of the time indices, as 
+well as the actual values of the time series, uncertain.
+
+To do this, we'll use the `example_uncertain_indexvalue_datasets` function that ships with 
+`UncertainData.jl`. It takes as input a `DiscreteDynamicalSystem` instance,
+the number of desired points, and which variables of the system to use for the 
+time series. Time series will be generated from a unidirectionally 
+coupled AR1 system from the [CausalityTools](https://github.com/kahaaga/CausalityTools.jl) 
+package. To simulate real-world data, some noise is added to the values and 
+indices.
+
+```julia
+using UncertainData, CausalityTools, Plots
+
+system = CausalityTools.ar1_unidir(c_xy = 0.5)
+vars = (1, 2)
+npts, tstep = 100, 10
+
+d_xind, d_yind = Uniform(2.5, 15.5), Uniform(2.5, 15.5)
+d_xval, d_yval = Uniform(0.01, 0.2), Uniform(0.01, 0.2)
+
+X, Y = example_uncertain_indexvalue_datasets(system,
+    npts, vars, tstep = tstep,
+    d_xind = d_xind, d_yind = d_yind,
+    d_xval = d_xval, d_yval = d_yval);
+```
+
+Let's plot the data.
+
+```julia 
+qs = [0.05, 0.95] # use the same quantile ranges for both indices and values
+plot(X, qs, qs, ms = 2, c = :black, marker = stroke(0.01, :black), 
+    xlabel = "Time step", ylabel = "Value")
+```
+
+![](imgs/example_timeseries_X.svg)
+
+Our data have uncertain time indices, so they are not on a regularly spaced grid.
+Let's say we want a grid where the left bin edges range from `0` to `1000` in 
+steps of `50`. Superimposed on our data, that grid looks as follows.
+
+```julia
+resampling = BinnedResampling(0:50:1000, 1000)
+
+qs = [0.05, 0.95] # plotting quantile ranges
+plot(X, qs, qs, ms = 2, c = :black, marker = stroke(0.01, :black), 
+    xlabel = "Time step", ylabel = "Value")
+vline!(0:50:1000 |> collect, label = "", c = :grey, lw = 0.5, ls = :dash)
+```
+
+![](imgs/example_timeseries_X_binning.svg)
+
+## `BinnedMeanResampling`
+
+Assume that the uncertainties in the time values are independent. Bin averages 
+can then be obtained by resampling every uncertain value in the dataset 
+many times, keeping track of which draws falls in which time bins, then taking
+the average of the draws in each of the bins. We'll resample each point 
+`10000` times. In total, the bin means are then computed based on 
+`100*10000` draws of the values in the dataset (we constructed the dataset 
+so that it has 100 points).
+
+```julia
+resampling = BinnedMeanResampling(0:50:1000, 10000)
[…]
