+++ docs/src/tutorials/tutorial_overview.md
@@ -1,0 +1,4 @@
+
+# Regularising uncertain data
+
+- [Transforming uncertain data to regular grid](@ref transform_data_to_regular_grid)
