### 2019-10-22 b6a6d4c
Improve `resample!` docs
