### 2019-10-22 ccdc8d0
Binned resampling, and in-place resampling
