+++ Project.toml
@@ -2,7 +2,7 @@ name = "UncertainData"
 uuid = "dcd9ba68-c27b-5cea-ae21-829cd07325bf"
 authors = ["Kristian Agasøster Haaga <kahaaga@gmail.com>"]
 repo = "https://github.com/kahaaga/UncertainData.jl.git"
-version = "0.7.0"
+version = "0.8.0"
 
 [deps]
 Bootstrap = "e28b5b4c-05e8-5b66-bc03-6f0c0a0a06e0"
@@ -24,7 +24,22 @@ StatsBase = "2913bbd2-ae8a-5f71-8c99-4fb6c76f3a91"
 Unitful = "1986cc42-f94f-5a68-af5c-568840ba703d"
 
 [compat]
-julia = ">= 1.1"
+julia = "^1.1"
+Distributions = "0.21, 1"
+HypothesisTests = "0.8, 1"
+DynamicalSystemsBase = "^1.3"
+Interpolations = "^0.12, ^1"
+Measurements = "^2.1"
+Bootstrap = "^2.2"
+Combinatorics = "^0.7.0, ^1"
+KernelDensity = "^0.5.0, ^1"
+StatsBase = "^0.32.0, ^1"
+DocumenterTools = "^0.1, ^1"
+Unitful = "^0.17.0, ^1"
+StaticArrays = "^0.11.0, ^1"
+IntervalArithmetic = "^0.16, ^1"
+RecipesBase = "^0.7, ^1"
+Reexport = "^0.2, ^1"
 
 [extras]
 Test = "8dfed614-e22c-5e08-85e1-65c5234f0b40"
