### 2019-10-22 5a46880
Add tests for BinnedMeanResampling and BinnedResampling
