+++ test/resampling/resampling_schemes/test_BinnedResampling.jl
@@ -1,0 +1,5 @@
+n_draws = 100
+grid = 0:10:500
+resampling = BinnedResampling(grid, n_draws)
+
+@test resampling isa BinnedResampling
