### 2019-10-22 6d13c5d
resample tuple of uncertain values into pre-allocated vector
