### 2019-10-22 69d657a
Update changelog and update message
