### 2019-10-23 bca2287
Improve docs for BinnedMeanWeightedResampling
