### 2019-10-23 f6a965c
Added `resample` method for `BinnedWeightedMeanResampling`
