### 2019-10-23 66a80be
Release 0.8.0 (#113)
* Improve docs and add tutorial on resampling with binning

* Improve `resample!` docs

* Binned resampling, and in-place resampling

* Update Project.toml

* Add tests for BinnedMeanResampling and BinnedResampling

* resample tuple of uncertain values into pre-allocated vector

* Update changelog

* Update changelog and update message

* Improve docs for BinnedMeanWeightedResampling

* Added `resample` method for `BinnedWeightedMeanResampling`
