### 2019-10-23 4ed3722
Add `rand(x::CertainValue, n::Int)` method
Plus tests
