### 2019-10-23 3ff44ef
Specify `rand(v::CertainValue, n::Int)`
