### 2019-10-23 c712d4e
Added constructors for vectors of numbers
