+++ test/uncertain_datasets/test_uncertain_index_dataset.jl
@@ -1,0 +1,1 @@
+@test UncertainIndexDataset(rand(30)) isa UncertainIndexDataset
