### 2019-10-23 9f43ffc
Added constructors for vectors of numbers
