+++ test/uncertain_datasets/test_uncertain_value_dataset.jl
@@ -1,0 +1,1 @@
+@test UncertainValueDataset(rand(30)) isa UncertainValueDataset
