### 2019-10-23 5a0f39a
Release 0.8.1 (#115)
# Release 0.8.1

## Bug fixes 

- Added `rand(x::CertainValue, n::Int)` method
- Added uncertain dataset constructors for vectors of numbers
