+++ test/resampling/uncertain_values/test_resampling_certain_value.jl
@@ -13,6 +13,9 @@ test_constraints = [
 
 T = eltype(x)
 
+@test rand(x) isa T
+@test rand(x, 10) isa Vector
+
 @test resample(x) isa T
 @test resample(x, 10) isa Vector
 @test all(resample(x, 10) .== x.value)
