### 2019-10-23 b040047
Added `resample` method for `BinnedWeightedResampling` scheme
