+++ src/resampling/resampling_with_schemes/resampling_schemes_binned.jl
@@ -33,11 +33,12 @@ d_xval = d_xval, d_yval = d_yval);
 
 n_draws = 10000 # draws per uncertain value
 time_grid = 0:50:1000
+wts = Weights(rand(length(X))) # some random weights
 
 # Resample both X and Y so that they are both at the same time indices, 
 # and take the mean of each bin.
-resampled_dataset = resample(X, BinnedMeanResampling(time_grid, n_draws))
-resampled_dataset = resample(Y, BinnedMeanResampling(time_grid, n_draws))
+resampled_dataset = resample(X, BinnedMeanWeightedResampling(time_grid, wts, n_draws))
+resampled_dataset = resample(Y, BinnedMeanWeightedResampling(time_grid, wts, n_draws))
 ```
 """
 function resample(x::AbstractUncertainIndexValueDataset, resampling::BinnedMeanWeightedResampling)
@@ -176,54 +177,6 @@ function resample(x::AbstractUncertainIndexValueDataset, resampling::BinnedMeanR
     return bin_avgs
 end
 
-function resample(x::AbstractUncertainIndexValueDataset, resampling::BinnedMeanWeightedResampling)
-     
-    # Represent entire dataset as a weighted population and sample from that 
-    pop_inds = UncertainValue(x.indices.indices, resampling.weights)
-    pop_vals = UncertainValue(x.values.values, resampling.weights)
-    
-    # Pre-allocate an array representing each bin, and an array keeping track
-    # of the values falling in that bin.
-    n_bins = length(resampling.left_bin_edges) - 1
-    bin_sums = fill(0.0, n_bins)
-    bin_sums_n_entries = similar(bin_sums)
-
-    # Pre-allocate some arrays into which we resample the values of the 
-    # index and value populations.
-    idxs = fill(NaN, resampling.n)
-    vals = fill(NaN, resampling.n)
-    
-    # Used to compute the index of the bin into which a draw belongs
-    mini = minimum(resampling.left_bin_edges)
-    s = step(resampling.left_bin_edges)
-    
-    for (j, (pop_idx, pop_val)) in enumerate(zip(pop_inds, pop_vals))
-    
-        # Sample the j-th idx-value pair `resampling.n` times and 
-        # accumulate the values in the correct bin sum. Also keep
-        # track of how many values there are in each bin.
-        resample!(idxs, pop_idx) 
-        resample!(vals, pop_val)
-        
-        @inbounds for i = 1:resampling.n
-            arr_idx = ceil(Int, (idxs[i] - mini) / s)
-            
-            # Because the indices of `x` are uncertain values 
-            # with potentially infinite support, we need to check
-            # that the value falls inside the grid
-            if 0 < arr_idx <= n_bins
-                bin_sums[arr_idx] += vals[i]
-                bin_sums_n_entries[arr_idx] += 1.0
-            end
-        end
-    end
-    
-    # Return bin averages (entries with 0s are represented as NaNs)
-    bin_avgs = bin_sums ./ bin_sums_n_entries
-    bin_avgs[isapprox.(bin_avgs, 0.0)] .= NaN
-    return bin_avgs
-end
-
 """
     resample(x::AbstractUncertainIndexValueDataset, resampling::BinnedResampling;
         nan_threshold = 0.0)
@@ -319,8 +272,110 @@ function resample(x::AbstractUncertainIndexValueDataset, resampling::BinnedResam
         end
     end
     
-    new_inds = UncertainIndexDataset(UncertainValue.(bin_edges .+ step(bin_edges)/2))
[…]
