### 2019-10-23 ffec3e2
Added tests for BinnedWeightedResampling
