### 2019-10-23 a6ff183
Added tests for BinnedWeightedMeanResampling
