+++ test/runtests.jl
@@ -40,6 +40,7 @@ include("resampling/test_resampling_uncertain_tuples.jl")
 include("resampling/resampling_schemes/test_BinnedResampling.jl")
 include("resampling/resampling_schemes/test_BinnedMeanResampling.jl")
 include("resampling/resampling_schemes/test_BinnedWeightedResampling.jl")
+include("resampling/resampling_schemes/test_BinnedWeightedMeanResampling.jl")
 
 include("resampling/resampling_schemes/test_ConstrainedIndexValueResampling.jl")
 include("resampling/resampling_schemes/test_ConstrainedValueResampling.jl")
