+++ test/resampling/resampling_schemes/test_BinnedWeightedResampling.jl
@@ -3,4 +3,4 @@ grid = 0:10:500
 wts = Weights(rand(10))
 resampling = BinnedWeightedResampling(grid, wts, n_draws)
 
-@test resampling isa BinnedResampling
+@test resampling isa BinnedWeightedResampling
