### 2019-10-24 e390260
Fix test path and update changelog
