### 2019-10-24 e4b266f
Add more specific abstract types for binned resamplings
