### 2019-10-24 e1f5aaf
Abstract resampling types
