+++ test/resampling/resampling_schemes/test_BinnedWeightedResampling.jl
@@ -4,4 +4,4 @@ wts = Weights(rand(10))
 resampling = BinnedWeightedResampling(grid, wts, n_draws)
 
 @test resampling isa BinnedWeightedResampling
-@test resampling isa AbstractBinnedResampling
+@test typeof(resampling) <: AbstractBinnedUncertainValueResampling
