### 2019-10-24 57b2abe
Release 0.8.2 (#117)
## Uncertaindata.jl v0.8.2

### New features

- Added `resample` method for `BinnedWeightedResampling` scheme.
- Added `AbstractBinnedResampling` for binned resamplings.
- Added `AbstractBinnedUncertainValueResampling` abstract type for binnings where 
    the values in each bin is represented by an uncertain value. `BinnedResampling` 
    and `BinnedWeightedResampling` are subtypes `AbstractBinnedUncertainValueResampling`.
- Added `AbstractBinnedSummarisedResampling` abstract type for binnings where the values 
    in each bin are summarised to a single value. `BinnedMean
[…]
