+++ test/runtests.jl
@@ -38,7 +38,10 @@ include("resampling/test_resampling_uncertain_tuples.jl")
 
 # Resampling schemes
 include("resampling/resampling_schemes/test_BinnedResampling.jl")
+include("resampling/resampling_schemes/test_BinnedWeightedResampling.jl")
 include("resampling/resampling_schemes/test_BinnedMeanResampling.jl")
+include("resampling/resampling_schemes/test_BinnedMeanWeightedResampling.jl")
+
 include("resampling/resampling_schemes/test_ConstrainedIndexValueResampling.jl")
 include("resampling/resampling_schemes/test_ConstrainedValueResampling.jl")
 include("resampling/resampling_schemes/test_SequentialResampling.jl")
