### 2019-10-24 #1666 @kthyng
@oliviaguest Having all the boxes checked is preferable, but it is clear this has been taken care of so the review can move forward. Do you want to go ahead and update the version number? Looks like it has continued to be updated in the meantime.

@kahaaga Which version of your package do you intend to be associated with the JOSS paper?
