### 2019-10-24 #1666 @kahaaga
The latest release of the package is v0.8.2 and the functionality is still the same (but with some additional minor features), so I guess the latest version could be associated with the paper?
