### 2019-10-25 #1666 @labarba
Authors need to make a tagged release of their software to include the latest revisions. The version number will thus increase. You, as handling editor, run `@whedon set v.xx as version`. That will update the version at the top of this issue, and then you are done, and can check off the item on the checklist.
