### 2019-10-29 0c95b76
Added `findall_nan_chunks` and `findall_nan_chunks!` methods
