### 2019-10-29 #1666 @kahaaga
@oliviaguest I've archived the software at Zenodo and added the DOI ([10.5281/zenodo.3522252](https://doi.org/10.5281/zenodo.3522252)) to the README . Should be good to go now!
