### 2019-10-30 #1666 @oliviaguest
@kahaaga can you check out that missing DOI thing and generally check it all looks fine still, please?
