### 2019-10-30 #1666 @kahaaga
@oliviaguest I've added the missing DOI now.

Also, in the final proof, one of the references in the first wasn't formatted correctly. I've tried to fix this by changing the citation from `[@author1, @author2]` to `[@author1; @author2]`, as per the documentation. Can we generate the proof again and check before depositing?
