### 2019-10-31 #1666 @arfon
☝️ @kahaaga - you can do this yourself with `@whedon generate pdf`
