### 2019-11-01 #1666 @danielskatz
@oliviaguest - please notify @openjournals/joss-eics when a submission is ready for acceptance (since we have a rotation of AEiCs).  Is that the case with this one?
